refactor(torghut): split runtime window import ownership
