+++ services/torghut/scripts/import_hypothesis_runtime_windows.py
@@ -49,204 +49,177 @@ from app.trading.runtime_window_import import (
     resolve_hypothesis_manifest,
 )
 
-EXECUTION_ELIGIBLE_DECISION_STATUSES = (
-    "submitted",
-    "filled",
-    "partially_filled",
+from scripts.hypothesis_runtime_window_import.common import (
+    _runtime_source_row_symbol as _runtime_source_row_symbol,
+    _execution_signed_qty as _execution_signed_qty,
+    _execution_row_has_fill as _execution_row_has_fill,
+    _source_authority_order_event_row as _source_authority_order_event_row,
+    _manifest_strategy_family_for_resolution as _manifest_strategy_family_for_resolution,
+    _parse_dt as _parse_dt,
+    _parse_dt_or_none as _parse_dt_or_none,
+    _as_mapping as _as_mapping,
+    _as_sequence as _as_sequence,
+    _parse_target_metadata as _parse_target_metadata,
+    _decimal_or_none as _decimal_or_none,
+    _text_or_none as _text_or_none,
+    _row_payloads as _row_payloads,
+    _first_decimal as _first_decimal,
+    _first_positive_decimal as _first_positive_decimal,
+    _first_text as _first_text,
+    _direct_text as _direct_text,
+    _bool_value as _bool_value,
+    _direct_bool as _direct_bool,
+    _text_values as _text_values,
+    _first_bool as _first_bool,
+    _position_snapshot_open_position_count as _position_snapshot_open_position_count,
+    _flat_start_position_snapshot_authority as _flat_start_position_snapshot_authority,
+    _runtime_ledger_equity_denominator_from_rows as _runtime_ledger_equity_denominator_from_rows,
+    _json_default as _json_default,
+    _stable_payload_digest as _stable_payload_digest,
+    _runtime_source_lineage_hash as _runtime_source_lineage_hash,
+    _attach_source_lineage_context as _attach_source_lineage_context,
+    _decimal_ceil_cent as _decimal_ceil_cent,
+    _row_has_alpaca_us_equity_order_source as _row_has_alpaca_us_equity_order_source,
+    _alpaca_2026_equity_fee_schedule_cost as _alpaca_2026_equity_fee_schedule_cost,
+    _alpaca_2026_equity_fee_schedule_hash as _alpaca_2026_equity_fee_schedule_hash,
+    _cost_basis_is_alpaca_fee_schedule as _cost_basis_is_alpaca_fee_schedule,
+    _first_payload_digest as _first_payload_digest,
+    _first_lineage_digest as _first_lineage_digest,
+    _merge_count_mappings as _merge_count_mappings,
+    _source_identifier_values as _source_identifier_values,
+    _source_offset_values as _source_offset_values,
+    _with_canonical_runtime_source_refs as _with_canonical_runtime_source_refs,
+    _nonnegative_int as _nonnegative_int,
+    _metadata_text_list as _metadata_text_list,
+    _runtime_ledger_tca_ref_texts as _runtime_ledger_tca_ref_texts,
+    _runtime_ledger_execution_tca_metric_refs as _runtime_ledger_execution_tca_metric_refs,
+    _metadata_symbol_list as _metadata_symbol_list,
+    _target_metadata_source_symbols as _target_metadata_source_symbols,
+    _runtime_ledger_target_metadata_artifact_refs as _runtime_ledger_target_metadata_artifact_refs,
+    _metadata_nonnegative_int_or_none as _metadata_nonnegative_int_or_none,
+    _runtime_window_source_kind_is_informational as _runtime_window_source_kind_is_informational,
+    _source_collection_target_authorization_blockers as _source_collection_target_authorization_blockers,
+    _source_kind_allows_runtime_ledger_materialization as _source_kind_allows_runtime_ledger_materialization,
+    _runtime_ledger_event_type as _runtime_ledger_event_type,
+    _runtime_ledger_row_time as _runtime_ledger_row_time,
+    _parse_artifact_window_datetime as _parse_artifact_window_datetime,
+    _window_weekday_count as _window_weekday_count,
+    _runtime_ledger_artifact_candidate_ids as _runtime_ledger_artifact_candidate_ids,
 )
-POST_COST_BASIS_RUNTIME_LEDGER = POST_COST_PNL_BASIS
-POST_COST_BASIS_EXECUTION_RECONSTRUCTION = (
-    "execution_reconstructed_pnl_after_explicit_costs"
+from scripts.hypothesis_runtime_window_import.constants import (
+    _LINEAGE_CONTEXT_KEYS as _LINEAGE_CONTEXT_KEYS,
+    _AUTHORITATIVE_RUNTIME_LEDGER_MATERIALIZATION_SOURCE_KINDS as _AUTHORITATIVE_RUNTIME_LEDGER_MATERIALIZATION_SOURCE_KINDS,
+    EXECUTION_ELIGIBLE_DECISION_STATUSES as EXECUTION_ELIGIBLE_DECISION_STATUSES,
+    POST_COST_BASIS_RUNTIME_LEDGER as POST_COST_BASIS_RUNTIME_LEDGER,
+    POST_COST_BASIS_EXECUTION_RECONSTRUCTION as POST_COST_BASIS_EXECUTION_RECONSTRUCTION,
+    RUNTIME_LEDGER_ARTIFACT_SCHEMAS as RUNTIME_LEDGER_ARTIFACT_SCHEMAS,
+    RUNTIME_LEDGER_BUCKET_SCHEMAS as RUNTIME_LEDGER_BUCKET_SCHEMAS,
+    EXACT_REPLAY_ARTIFACT_AUTHORITY_CLASS as EXACT_REPLAY_ARTIFACT_AUTHORITY_CLASS,
+    EXACT_REPLAY_ARTIFACT_AUTHORITY_BLOCKERS as EXACT_REPLAY_ARTIFACT_AUTHORITY_BLOCKERS,
+    RUNTIME_LEDGER_SOURCE_WINDOW_MISSING_BLOCKER as RUNTIME_LEDGER_SOURCE_WINDOW_MISSING_BLOCKER,
+    RUNTIME_LEDGER_CARRY_IN_LOOKBACK as RUNTIME_LEDGER_CARRY_IN_LOOKBACK,
+    RUNTIME_LEDGER_POST_WINDOW_CLOSEOUT_LOOKAHEAD as RUNTIME_LEDGER_POST_WINDOW_CLOSEOUT_LOOKAHEAD,
+    RUNTIME_LEDGER_FLAT_START_SNAPSHOT_STALE_SECONDS as RUNTIME_LEDGER_FLAT_START_SNAPSHOT_STALE_SECONDS,
+    RUNTIME_LEDGER_FLAT_START_SNAPSHOT_AFTER_START_GRACE_SECONDS as RUNTIME_LEDGER_FLAT_START_SNAPSHOT_AFTER_START_GRACE_SECONDS,
+    RUNTIME_LEDGER_SOURCE_REFS_MISSING_BLOCKER as RUNTIME_LEDGER_SOURCE_REFS_MISSING_BLOCKER,
+    RUNTIME_LEDGER_EXECUTION_TCA_REFS_MISSING_BLOCKER as RUNTIME_LEDGER_EXECUTION_TCA_REFS_MISSING_BLOCKER,
+    CONFIGURED_PAPER_COLLECTION_HYPOTHESIS_PREFIX as CONFIGURED_PAPER_COLLECTION_HYPOTHESIS_PREFIX,
+    CONFIGURED_SIMPLE_LANE_PAPER_SOURCE_KIND as CONFIGURED_SIMPLE_LANE_PAPER_SOURCE_KIND,
+    RUNTIME_LEDGER_AUTHORITY_CLASS_MISSING_BLOCKER as RUNTIME_LEDGER_AUTHORITY_CLASS_MISSING_BLOCKER,
+    EXECUTION_TCA_MISSING_BLOCKER as EXECUTION_TCA_MISSING_BLOCKER,
+    _RUNTIME_LEDGER_PROMOTION_GRADE_AUTHORITY_MARKERS as _RUNTIME_LEDGER_PROMOTION_GRADE_AUTHORITY_MARKERS,
+    _RUNTIME_LEDGER_DECISION_EVENTS as _RUNTIME_LEDGER_DECISION_EVENTS,
+    _RUNTIME_LEDGER_FILL_EVENTS as _RUNTIME_LEDGER_FILL_EVENTS,
+    _RUNTIME_LEDGER_ORDER_LIFECYCLE_EVENTS as _RUNTIME_LEDGER_ORDER_LIFECYCLE_EVENTS,
+    ORDER_FEED_FILL_LIFECYCLE_MISSING_BLOCKER as ORDER_FEED_FILL_LIFECYCLE_MISSING_BLOCKER,
+    ORDER_FEED_FILL_LIFECYCLE_INCOMPLETE_BLOCKER as ORDER_FEED_FILL_LIFECYCLE_INCOMPLETE_BLOCKER,
+    ORDER_FEED_FILL_LIFECYCLE_ORDER_ID_MISSING_BLOCKER as ORDER_FEED_FILL_LIFECYCLE_ORDER_ID_MISSING_BLOCKER,
+    ORDER_FEED_UNLINKED_FILL_LIFECYCLE_PRESENT_BLOCKER as ORDER_FEED_UNLINKED_FILL_LIFECYCLE_PRESENT_BLOCKER,
+    ORDER_FEED_LIFECYCLE_MISSING_BLOCKER as ORDER_FEED_LIFECYCLE_MISSING_BLOCKER,
+    EXECUTION_ECONOMICS_MISSING_BLOCKER as EXECUTION_ECONOMICS_MISSING_BLOCKER,
+    ORDER_FEED_FILL_DELTA_BASIS_MISSING_BLOCKER as ORDER_FEED_FILL_DELTA_BASIS_MISSING_BLOCKER,
+    ORDER_FEED_FILL_DELTA_MISSING_BLOCKER as ORDER_FEED_FILL_DELTA_MISSING_BLOCKER,
+    _EXECUTION_TCA_REF_KEYS as _EXECUTION_TCA_REF_KEYS,
+    _SOURCE_WINDOW_REF_KEYS as _SOURCE_WINDOW_REF_KEYS,
+    _EXECUTION_ORDER_EVENT_REF_KEYS as _EXECUTION_ORDER_EVENT_REF_KEYS,
+    _RUNTIME_LEDGER_SOURCE_AUTHORITY_BLOCKERS as _RUNTIME_LEDGER_SOURCE_AUTHORITY_BLOCKERS,
+    _RUNTIME_LIFECYCLE_IDENTIFIER_KEYS as _RUNTIME_LIFECYCLE_IDENTIFIER_KEYS,
+    ALPACA_2026_EQUITY_SEC_FEE_RATE as ALPACA_2026_EQUITY_SEC_FEE_RATE,
+    ALPACA_2026_EQUITY_TAF_RATE_PER_SHARE as ALPACA_2026_EQUITY_TAF_RATE_PER_SHARE,
+    ALPACA_2026_EQUITY_TAF_CAP as ALPACA_2026_EQUITY_TAF_CAP,
+    ALPACA_2026_FEE_SCHEDULE_REVISED_ON as ALPACA_2026_FEE_SCHEDULE_REVISED_ON,
+    _CENT as _CENT,
+    _RUNTIME_LEDGER_EQUITY_DENOMINATOR_KEYS as _RUNTIME_LEDGER_EQUITY_DENOMINATOR_KEYS,
+    SOURCE_LINEAGE_CANDIDATE_KEYS as SOURCE_LINEAGE_CANDIDATE_KEYS,
+    SOURCE_LINEAGE_HYPOTHESIS_KEYS as SOURCE_LINEAGE_HYPOTHESIS_KEYS,
+    SOURCE_DECISION_MODE_MISSING_PARTITION as SOURCE_DECISION_MODE_MISSING_PARTITION,
 )
-RUNTIME_LEDGER_ARTIFACT_SCHEMAS = frozenset(
-    {
-        EXACT_REPLAY_LEDGER_SCHEMA_VERSION,
-        "torghut.exact_replay_ledger.rows.v1",
-    }
+from scripts.hypothesis_runtime_window_import.runtime_ledger_authority import (
+    _promotion_grade_runtime_ledger_authority_marker_present as _promotion_grade_runtime_ledger_authority_marker_present,
+    runtime_ledger_promotion_source_authority_blockers as runtime_ledger_promotion_source_authority_blockers,
+    runtime_ledger_promotion_source_authority_present as runtime_ledger_promotion_source_authority_present,
+    _runtime_ledger_bucket_payload as _runtime_ledger_bucket_payload,
+    _mapping_hash_count as _mapping_hash_count,
+    _positive_count_mapping_present as _positive_count_mapping_present,
+    _runtime_ledger_explicit_costs_present as _runtime_ledger_explicit_costs_present,
+    _runtime_ledger_bucket_profit_proof_blockers as _runtime_ledger_bucket_profit_proof_blockers,
+    _runtime_ledger_bucket_profit_proof_present as _runtime_ledger_bucket_profit_proof_present,
+    _runtime_ledger_tca_row_from_bucket as _runtime_ledger_tca_row_from_bucket,
+    _append_runtime_ledger_tca_row_blocker as _append_runtime_ledger_tca_row_blocker,
+    _with_runtime_ledger_source_authority_context as _with_runtime_ledger_source_authority_context,
 )
-RUNTIME_LEDGER_BUCKET_SCHEMAS = frozenset(
-    {
-        EXACT_REPLAY_LEDGER_SCHEMA_VERSION,
-        "torghut.runtime-ledger-bucket.v1",
-    }
+from scripts.hypothesis_runtime_window_import.runtime_observation import (
+    _mark_runtime_ledger_tca_rows_as_exact_replay_artifacts as _mark_runtime_ledger_tca_rows_as_exact_replay_artifacts,
+    _runtime_ledger_target_metadata_blockers as _runtime_ledger_target_metadata_blockers,
+    _runtime_window_import_proof_hygiene_blockers as _runtime_window_import_proof_hygiene_blockers,
+    _runtime_ledger_profit_proof_present as _runtime_ledger_profit_proof_present,
+    _runtime_observation_authority_payload as _runtime_observation_authority_payload,
+    _runtime_ledger_tca_materialization_metadata as _runtime_ledger_tca_materialization_metadata,
+    _runtime_ledger_tca_rows_from_observed_buckets as _runtime_ledger_tca_rows_from_observed_buckets,
+    _runtime_observation_payload_with_observed_runtime_ledger_materialization as _runtime_observation_payload_with_observed_runtime_ledger_materialization,
+    _runtime_window_import_audit_summary as _runtime_window_import_audit_summary,
 )
-EXACT_REPLAY_ARTIFACT_AUTHORITY_CLASS = "exact_replay_artifact_only_not_live"
-EXACT_REPLAY_ARTIFACT_AUTHORITY_BLOCKERS = (
-    "exact_replay_artifact_not_runtime_proof",
-    "runtime_ledger_source_window_missing",
-    "runtime_ledger_source_refs_missing",
+from scripts.hypothesis_runtime_window_import.source_decisions import (
+    _source_decision_priority_payloads as _source_decision_priority_payloads,
+    _source_decision_mode as _source_decision_mode,
+    _source_decision_profit_proof_flag as _source_decision_profit_proof_flag,
+    _source_decision_target_notional_sizing_audit as _source_decision_target_notional_sizing_audit,
+    _source_decision_target_notional_sizing_summary as _source_decision_target_notional_sizing_summary,
+    _source_row_is_paper_route_probe_exit as _source_row_is_paper_route_probe_exit,
+    _source_decision_identifier_values as _source_decision_identifier_values,
+    _paper_route_probe_exit_identifiers as _paper_route_probe_exit_identifiers,
+    _source_row_is_paper_route_probe_exit_or_linked as _source_row_is_paper_route_probe_exit_or_linked,
+    _source_decision_mode_counts as _source_decision_mode_counts,
+    _source_decision_mode_lookup as _source_decision_mode_lookup,
+    _source_decision_mode_partition_key as _source_decision_mode_partition_key,
+    _partition_runtime_source_rows_by_decision_mode as _partition_runtime_source_rows_by_decision_mode,
+    _source_decision_rows_profit_proof_eligible as _source_decision_rows_profit_proof_eligible,
+    _source_row_matches_lineage as _source_row_matches_lineage,
+    _source_row_lineage_missing_or_matches as _source_row_lineage_missing_or_matches,
+    _source_row_time_before as _source_row_time_before,
+    _source_row_time_after_window as _source_row_time_after_window,
+    _runtime_open_qtys_by_symbol_from_execution_rows as _runtime_open_qtys_by_symbol_from_execution_rows,
+    _source_decision_action_offsets_open_qty as _source_decision_action_offsets_open_qty,
+    _source_row_decision_ids as _source_row_decision_ids,
+    _source_row_execution_ids as _source_row_execution_ids,
+    _mark_post_window_closeout_source_row as _mark_post_window_closeout_source_row,
+    _filter_source_rows_for_runtime_window as _filter_source_rows_for_runtime_window,
 )
-RUNTIME_LEDGER_SOURCE_WINDOW_MISSING_BLOCKER = "runtime_ledger_source_window_missing"
-RUNTIME_LEDGER_CARRY_IN_LOOKBACK = timedelta(days=5)
-RUNTIME_LEDGER_POST_WINDOW_CLOSEOUT_LOOKAHEAD = timedelta(seconds=3600)
-RUNTIME_LEDGER_FLAT_START_SNAPSHOT_STALE_SECONDS = 900
-RUNTIME_LEDGER_FLAT_START_SNAPSHOT_AFTER_START_GRACE_SECONDS = 300
-RUNTIME_LEDGER_SOURCE_REFS_MISSING_BLOCKER = "runtime_ledger_source_refs_missing"
-RUNTIME_LEDGER_EXECUTION_TCA_REFS_MISSING_BLOCKER = (
-    "runtime_ledger_execution_tca_refs_missing"
+from scripts.hypothesis_runtime_window_import.source_windows import (
+    _unique_source_window_rows as _unique_source_window_rows,
+    _source_window_status_counts as _source_window_status_counts,
+    _source_window_classification_counts as _source_window_classification_counts,
+    _source_window_gap_count as _source_window_gap_count,
+    _source_window_gap_ranges as _source_window_gap_ranges,
+    _decision_lifecycle_query_row as _decision_lifecycle_query_row,
+    _decision_lifecycle_query_row_has_time as _decision_lifecycle_query_row_has_time,
+    _execution_query_row as _execution_query_row,
+    _execution_query_row_has_time as _execution_query_row_has_time,
+    _execution_query_row_has_tca_ref as _execution_query_row_has_tca_ref,
+    _source_window_query_context as _source_window_query_context,
+    _order_lifecycle_query_row as _order_lifecycle_query_row,
+    _order_lifecycle_query_row_has_time as _order_lifecycle_query_row_has_time,
 )
-CONFIGURED_PAPER_COLLECTION_HYPOTHESIS_PREFIX = "configured-paper-collection:"
-CONFIGURED_SIMPLE_LANE_PAPER_SOURCE_KIND = (
-    "configured_simple_lane_paper_data_collection"
-)
-
-
-def _manifest_strategy_family_for_resolution(
-    *,
-    hypothesis_id: str,
-    strategy_family: str,
-    source_kind: str,
-    source_manifest_ref: str,
-) -> str | None:
-    normalized_strategy_family = strategy_family.strip() or None
-    if (
-        hypothesis_id.startswith(CONFIGURED_PAPER_COLLECTION_HYPOTHESIS_PREFIX)
-        and source_kind == CONFIGURED_SIMPLE_LANE_PAPER_SOURCE_KIND
-        and source_manifest_ref.strip()
-    ):
-        return None
-    return normalized_strategy_family
-
-
-RUNTIME_LEDGER_AUTHORITY_CLASS_MISSING_BLOCKER = (
-    "runtime_ledger_authority_class_missing"
-)
-EXECUTION_TCA_MISSING_BLOCKER = "execution_tca_missing"
-_RUNTIME_LEDGER_PROMOTION_GRADE_AUTHORITY_MARKERS = frozenset(
-    {
-        "runtime_order_feed_execution_source",
-        "event_sourced_runtime_ledger_profit_proof",
-        "source_execution_runtime_ledger_materialized",
-        "execution_order_events_runtime_ledger",
-        "source_execution_lifecycle_materialized_runtime_ledger",
-    }
-)
-_RUNTIME_LEDGER_DECISION_EVENTS = frozenset(
-    {"decision", "trade_decision", "signal_decision"}
-)
-_RUNTIME_LEDGER_FILL_EVENTS = frozenset(
-    {"fill", "filled", "partial_fill", "partially_filled"}
-)
-_RUNTIME_LEDGER_ORDER_LIFECYCLE_EVENTS = _RUNTIME_LEDGER_FILL_EVENTS | frozenset(
-    {"order_submitted"}
-)
-ORDER_FEED_FILL_LIFECYCLE_MISSING_BLOCKER = "order_feed_fill_lifecycle_missing"
-ORDER_FEED_FILL_LIFECYCLE_INCOMPLETE_BLOCKER = "order_feed_fill_lifecycle_incomplete"
-ORDER_FEED_FILL_LIFECYCLE_ORDER_ID_MISSING_BLOCKER = (
-    "order_feed_fill_lifecycle_order_id_missing"
-)
-ORDER_FEED_UNLINKED_FILL_LIFECYCLE_PRESENT_BLOCKER = (
-    "order_feed_unlinked_fill_lifecycle_present"
-)
-ORDER_FEED_LIFECYCLE_MISSING_BLOCKER = "order_feed_lifecycle_missing"
-EXECUTION_ECONOMICS_MISSING_BLOCKER = "execution_economics_missing"
-ORDER_FEED_FILL_DELTA_BASIS_MISSING_BLOCKER = "order_feed_fill_delta_basis_missing"
-ORDER_FEED_FILL_DELTA_MISSING_BLOCKER = "order_feed_fill_delta_missing"
-_EXECUTION_TCA_REF_KEYS = (
-    "execution_tca_metric_ids",
-    "execution_tca_metric_refs",
-    "execution_tca_metric_id",
-    "execution_tca_metric_ref",
-    "execution_tca_metrics",
-    "execution_tca_ids",
-    "execution_tca_refs",
-    "execution_tca_id",
-    "execution_tca_ref",
-    "runtime_ledger_execution_tca_metric_ids",
-    "runtime_ledger_execution_tca_metric_refs",
-    "runtime_ledger_execution_tca_metric_id",
-    "runtime_ledger_execution_tca_metric_ref",
-    "runtime_ledger_execution_tca_ids",
-    "runtime_ledger_execution_tca_refs",
-    "runtime_ledger_execution_tca_id",
-    "runtime_ledger_execution_tca_ref",
-    "tca_metric_ids",
-    "tca_metric_refs",
-    "tca_metric_id",
-    "tca_metric_ref",
-    "tca_ids",
-    "tca_refs",
-    "tca_id",
-    "tca_ref",
-)
-_SOURCE_WINDOW_REF_KEYS = (
-    "source_window_ids",
-    "source_window_refs",
-    "source_window_id",
-    "source_window_ref",
-    "runtime_ledger_source_window_ids",
-    "runtime_ledger_source_window_refs",
-    "runtime_ledger_source_window_id",
-    "runtime_ledger_source_window_ref",
-)
-_EXECUTION_ORDER_EVENT_REF_KEYS = (
-    "execution_order_event_ids",
-    "execution_order_event_refs",
-    "execution_order_event_id",
-    "execution_order_event_ref",
-    "runtime_ledger_execution_order_event_ids",
-    "runtime_ledger_execution_order_event_refs",
-    "runtime_ledger_execution_order_event_id",
-    "runtime_ledger_execution_order_event_ref",
-)
-_RUNTIME_LEDGER_SOURCE_AUTHORITY_BLOCKERS = frozenset(
-    {
-        "runtime_ledger_source_window_missing",
-        "runtime_ledger_source_window_ids_missing",
-        "runtime_ledger_source_refs_missing",
-        "runtime_ledger_trade_decision_refs_missing",
-        "runtime_ledger_execution_refs_missing",
-        "runtime_ledger_execution_order_event_refs_missing",
-        "runtime_ledger_source_offsets_missing",
-        "runtime_ledger_source_materialization_missing",
-        "runtime_ledger_authority_class_missing",
-        "order_feed_source_window_gap",
-        "order_feed_lifecycle_missing",
-        "execution_economics_missing",
-    }
-)
-_RUNTIME_LIFECYCLE_IDENTIFIER_KEYS = (
-    "execution_id",
-    "trade_decision_id",
-    "decision_id",
-    "decision_hash",
-    "order_id",
-    "alpaca_order_id",
-    "client_order_id",
-    "execution_correlation_id",
-)
-ALPACA_2026_EQUITY_SEC_FEE_RATE = Decimal("20.60") / Decimal("1000000")
-ALPACA_2026_EQUITY_TAF_RATE_PER_SHARE = Decimal("0.000195")
-ALPACA_2026_EQUITY_TAF_CAP = Decimal("9.79")
-ALPACA_2026_FEE_SCHEDULE_REVISED_ON = "2026-04-01"
-_CENT = Decimal("0.01")
-_RUNTIME_LEDGER_EQUITY_DENOMINATOR_KEYS = (
-    "account_equity",
-    "portfolio_equity",
-    "start_equity",
-    "starting_equity",
-    "equity",
-    "portfolio_value",
-    "net_liquidation",
-    "net_liquidation_value",
-)
-SOURCE_LINEAGE_CANDIDATE_KEYS = (
-    "candidate_id",
-    "candidate_ids",
-    "strategy_candidate_id",
-    "strategy_candidate_ids",
-    "source_candidate_id",
-    "source_candidate_ids",
-)
-SOURCE_LINEAGE_HYPOTHESIS_KEYS = (
-    "hypothesis_id",
-    "hypothesis_ids",
-    "strategy_hypothesis_id",
-    "strategy_hypothesis_ids",
-    "source_hypothesis_id",
-    "source_hypothesis_ids",
-)
-SOURCE_DECISION_MODE_MISSING_PARTITION = "source_decision_mode_missing"
 
 
 def _parse_args() -> argparse.Namespace:
@@ -358,2844 +331,6 @@ def _flag(value: str) -> bool:
     return value.strip().lower() not in {"", "0", "false", "no", "off"}
 
 
-def _parse_dt(value: str) -> datetime:
-    parsed = datetime.fromisoformat(value.replace("Z", "+00:00"))
-    if parsed.tzinfo is None:
-        return parsed.replace(tzinfo=timezone.utc)
-    return parsed.astimezone(timezone.utc)
-
-
-def _parse_dt_or_none(value: Any) -> datetime | None:
-    if isinstance(value, datetime):
-        return (
-            value.astimezone(timezone.utc)
-            if value.tzinfo
-            else value.replace(tzinfo=timezone.utc)
-        )
-    text = str(value or "").strip()
-    if not text:
-        return None
-    try:
-        return _parse_dt(text)
-    except Exception:
-        return None
-
-
-def _as_mapping(value: Any) -> dict[str, Any]:
-    return (
-        {str(key): item for key, item in value.items()}
-        if isinstance(value, Mapping)
-        else {}
-    )
-
-
-def _as_sequence(value: Any) -> list[Any] | None:
-    if value is None:
-        return []
-    if isinstance(value, str):
-        try:
-            parsed = json.loads(value)
-        except Exception:
-            return None
-        return _as_sequence(parsed)
-    if isinstance(value, Sequence) and not isinstance(value, (bytes, bytearray)):
-        return list(value)
-    return None
-
-
-def _parse_target_metadata(raw: str) -> dict[str, Any]:
-    text = str(raw or "").strip()
-    if not text:
-        return {}
-    try:
-        payload = json.loads(text)
-    except Exception as exc:
-        raise RuntimeError("target_metadata_json_invalid") from exc
-    if not isinstance(payload, Mapping):
-        raise RuntimeError("target_metadata_json_not_mapping")
-    return {str(key): value for key, value in payload.items()}
-
-
-def _decimal_or_none(value: Any) -> Decimal | None:
-    if value is None:
-        return None
-    text = str(value).strip()
-    if not text:
-        return None
-    try:
-        return Decimal(text)
-    except Exception:
-        return None
-
-
-def _text_or_none(value: Any) -> str | None:
-    text = str(value or "").strip()
-    return text or None
-
-
-def _promotion_grade_runtime_ledger_authority_marker_present(
-    bucket: Mapping[str, object],
-    key: str,
-) -> bool:
-    marker = _text_or_none(bucket.get(key))
-    return (
-        marker is not None
-        and marker in _RUNTIME_LEDGER_PROMOTION_GRADE_AUTHORITY_MARKERS
-    )
-
-
-def runtime_ledger_promotion_source_authority_blockers(
-    bucket: Mapping[str, object],
-) -> list[str]:
-    blockers = _base_runtime_ledger_promotion_source_authority_blockers(bucket)
-    if not (
-        _promotion_grade_runtime_ledger_authority_marker_present(
-            bucket, "authority_class"
-        )
-        and _promotion_grade_runtime_ledger_authority_marker_present(
-            bucket, "authority_reason"
-        )
-    ):
-        blockers.append(RUNTIME_LEDGER_AUTHORITY_CLASS_MISSING_BLOCKER)
-    return list(dict.fromkeys(blockers))
-
-
-def runtime_ledger_promotion_source_authority_present(
-    bucket: Mapping[str, object],
-) -> bool:
-    return not runtime_ledger_promotion_source_authority_blockers(bucket)
-
-
-def _row_payloads(row: Mapping[str, object]) -> list[Mapping[str, object]]:
-    payloads: list[Mapping[str, object]] = []
-
-    def append_payload(value: object, *, depth: int) -> None:
-        if not isinstance(value, Mapping):
-            return
-        payload = {str(item_key): item for item_key, item in value.items()}
-        payloads.append(payload)
-        if depth <= 0:
-            return
-        for nested in payload.values():
-            if isinstance(nested, Mapping):
-                append_payload(nested, depth=depth - 1)
-
-    append_payload(row, depth=4)
-    return payloads
-
-
-def _first_decimal(row: Mapping[str, object], *keys: str) -> Decimal | None:
-    for payload in _row_payloads(row):
-        for key in keys:
-            value = payload.get(key)
-            if (parsed := _decimal_or_none(value)) is not None:
-                return parsed
-    return None
-
-
-def _first_positive_decimal(row: Mapping[str, object], *keys: str) -> Decimal | None:
-    for payload in _row_payloads(row):
-        for key in keys:
-            value = payload.get(key)
-            if (parsed := _decimal_or_none(value)) is not None and parsed > 0:
-                return parsed
-    return None
-
-
-def _first_text(row: Mapping[str, object], *keys: str) -> str | None:
-    for payload in _row_payloads(row):
-        for key in keys:
-            text = str(payload.get(key) or "").strip()
-            if text:
-                return text
-    return None
-
-
-def _direct_text(row: Mapping[str, object], *keys: str) -> str | None:
-    for key in keys:
-        text = str(row.get(key) or "").strip()
-        if text:
-            return text
-    return None
-
-
-def _bool_value(value: object) -> bool | None:
-    if isinstance(value, bool):
-        return value
-    if isinstance(value, int | float | Decimal):
-        return bool(value)
-    text = str(value or "").strip().lower()
-    if text in {"1", "true", "yes", "on", "pass", "passed", "ok"}:
-        return True
-    if text in {"0", "false", "no", "off", "fail", "failed", "blocked"}:
-        return False
-    return None
-
-
-def _direct_bool(row: Mapping[str, object], *keys: str) -> bool | None:
-    for key in keys:
-        value = row.get(key)
-        if (parsed := _bool_value(value)) is not None:
-            return parsed
-    return None
-
-
-def _text_values(row: Mapping[str, object], *keys: str) -> set[str]:
-    values: set[str] = set()
-    for payload in _row_payloads(row):
-        for key in keys:
-            raw_value = payload.get(key)
-            if isinstance(raw_value, Sequence) and not isinstance(
-                raw_value, (str, bytes, bytearray)
-            ):
-                for item in raw_value:
-                    text = str(item or "").strip()
-                    if text:
-                        values.add(text)
-            else:
-                text = str(raw_value or "").strip()
-                if text:
-                    values.add(text)
-    return values
-
-
-def _first_bool(row: Mapping[str, object], *keys: str) -> bool | None:
-    for payload in _row_payloads(row):
-        for key in keys:
-            if (parsed := _bool_value(payload.get(key))) is not None:
-                return parsed
-    return None
-
-
-def _position_snapshot_open_position_count(positions: object) -> int | None:
-    position_rows = _as_sequence(positions)
-    if position_rows is None:
-        return None
-    count = 0
-    for raw_position in position_rows:
-        position = _as_mapping(raw_position)
-        if not position:
-            return None
-        qty = _decimal_or_none(
-            position.get("qty")
-            or position.get("quantity")
-            or position.get("filled_qty")
-            or position.get("position_qty")
-        )
-        if qty is None:
-            return None
-        if qty != 0:
-            count += 1
-    return count
-
-
-def _flat_start_position_snapshot_authority(
-    snapshot: Mapping[str, object] | None,
-) -> dict[str, object] | None:
-    if not snapshot:
-        return None
-    snapshot_id = _text_or_none(
-        snapshot.get("snapshot_id")
-        or snapshot.get("position_snapshot_id")
-        or snapshot.get("id")
-    )
-    snapshot_as_of = _text_or_none(
-        snapshot.get("snapshot_as_of")
-        or snapshot.get("position_snapshot_as_of")
-        or snapshot.get("as_of")
-    )
-    if snapshot_id is None or snapshot_as_of is None:
-        return None
-    if _metadata_text_list(snapshot.get("blockers")):
-        return None
-    explicit_flat = _first_bool(
-        snapshot, "flat", "account_flat", "zero_open_position_evidence"
-    )
-    positions = snapshot.get("positions")
-    position_count = _nonnegative_int(snapshot.get("position_count"))
-    if positions is not None:
-        position_count = _position_snapshot_open_position_count(positions)
-        if position_count is None:
-            return None
-    if explicit_flat is not True or position_count != 0:
-        return None
-    return {
-        "flat_start_position_snapshot_id": snapshot_id,
-        "flat_start_position_snapshot_as_of": snapshot_as_of,
-        "flat_start_position_snapshot_source": _text_or_none(
-            snapshot.get("snapshot_source") or snapshot.get("source")
-        )
-        or "position_snapshots",
-        "flat_start_position_snapshot_scope": _text_or_none(snapshot.get("scope"))
-        or "account_position_snapshot_at_runtime_window_start",
-        "carry_in_rows_suppressed_by_flat_start_snapshot": True,
-    }
-
-
-def _source_decision_priority_payloads(
-    row: Mapping[str, object],
-) -> list[Mapping[str, object]]:
-    payloads: list[Mapping[str, object]] = [row]
-    decision_json = _as_mapping(row.get("decision_json"))
-    if decision_json:
-        payloads.append(decision_json)
-        params = _as_mapping(decision_json.get("params"))
-        if params:
-            payloads.append(params)
-        for key in (
-            "strategy_signal_paper",
-            "paper_route_target_plan_source_decision",
-            "paper_route_target_plan",
-            "paper_route_target",
-        ):
-            if nested := _as_mapping(decision_json.get(key)):
-                payloads.append(nested)
-        if params:
-            for key in (
-                "strategy_signal_paper",
-                "paper_route_target_plan_source_decision",
-                "paper_route_target_plan",
-                "paper_route_target",
-            ):
-                if nested := _as_mapping(params.get(key)):
-                    payloads.append(nested)
-    return payloads
-
-
-def _source_decision_mode(row: Mapping[str, object]) -> str | None:
-    for payload in _source_decision_priority_payloads(row):
-        explicit = normalize_source_decision_mode(
-            _direct_text(payload, "source_decision_mode")
-        )
-        if explicit is not None:
-            return explicit
-    explicit = normalize_source_decision_mode(_first_text(row, "source_decision_mode"))
-    if explicit is not None:
-        return explicit
-    return normalize_source_decision_mode(row.get("mode"))
-
-
-def _source_decision_profit_proof_flag(row: Mapping[str, object]) -> bool | None:
-    for payload in _source_decision_priority_payloads(row):
-        value = _direct_bool(
-            payload,
-            "profit_proof_eligible",
-            "post_cost_promotion_eligible",
-        )
-        if value is not None:
-            return value
-    return _first_bool(row, "profit_proof_eligible", "post_cost_promotion_eligible")
-
-
-def _source_decision_target_notional_sizing_audit(
-    row: Mapping[str, object],
-) -> dict[str, object] | None:
-    for payload in _source_decision_priority_payloads(row):
-        audit = _as_mapping(payload.get("paper_route_target_notional_sizing"))
-        if audit:
-            return {str(key): value for key, value in audit.items()}
-        simple_lane = _as_mapping(payload.get("simple_lane"))
-        audit = _as_mapping(simple_lane.get("paper_route_target_notional_sizing"))
-        if audit:
-            return {str(key): value for key, value in audit.items()}
-    return None
-
-
-def _source_decision_target_notional_sizing_summary(
-    rows: Sequence[Mapping[str, object]],
-) -> dict[str, object]:
-    source_row_count = len(rows)
-    paper_route_probe_exit_identifiers = _paper_route_probe_exit_identifiers(rows)
-    sizing_required_rows = [
-        row
-        for row in rows
-        if not _source_row_is_paper_route_probe_exit_or_linked(
-            row,
-            paper_route_probe_exit_identifiers=paper_route_probe_exit_identifiers,
-        )
-    ]
-    sizing_required_row_count = len(sizing_required_rows)
-    audits: list[dict[str, object]] = []
-    authoritative_count = 0
-    non_authoritative_source_counts: dict[str, int] = {}
-    blocker_counts: dict[str, int] = {}
-    for row in sizing_required_rows:
-        audit = _source_decision_target_notional_sizing_audit(row)
-        if audit is None:
-            continue
-        audits.append(audit)
-        sizing_source = _text_or_none(audit.get("sizing_source")) or "missing"
-        if sizing_source == "target_notional":
-            authoritative_count += 1
-        else:
-            non_authoritative_source_counts[sizing_source] = (
-                non_authoritative_source_counts.get(sizing_source, 0) + 1
-            )
-        for blocker in _metadata_text_list(audit.get("blockers")):
-            blocker_counts[blocker] = blocker_counts.get(blocker, 0) + 1
-
-    mode_counts = _source_decision_mode_counts(rows)
-    requires_target_notional_sizing = (
-        bool(audits)
-        or mode_counts.get("bounded_paper_route_collection", 0) > 0
-        or any(
-            _first_text(row, "paper_route_probe_target_notional", "target_notional")
-            for row in sizing_required_rows
-        )
-    )
-    missing_count = max(sizing_required_row_count - len(audits), 0)
-    blockers: list[str] = []
-    if requires_target_notional_sizing and missing_count:
-        blockers.append("paper_route_target_notional_sizing_missing")
-    if non_authoritative_source_counts:
-        blockers.append("paper_route_target_notional_sizing_not_authoritative")
-    blockers.extend(blocker_counts)
-    return {
-        "source_row_count": source_row_count,
-        "target_notional_sizing_required_source_row_count": (sizing_required_row_count),
-        "excluded_paper_route_probe_exit_row_count": (
-            source_row_count - sizing_required_row_count
-        ),
-        "audit_count": len(audits),
-        "authoritative_target_notional_sizing_count": authoritative_count,
-        "missing_target_notional_sizing_count": (
-            missing_count if requires_target_notional_sizing else 0
-        ),
-        "non_authoritative_sizing_source_counts": dict(
-            sorted(non_authoritative_source_counts.items())
-        ),
-        "sizing_blocker_counts": dict(sorted(blocker_counts.items())),
-        "requires_target_notional_sizing": requires_target_notional_sizing,
-        "blockers": list(dict.fromkeys(blockers)),
-    }
-
-
-def _source_row_is_paper_route_probe_exit(row: Mapping[str, object]) -> bool:
-    if _first_bool(row, "post_window_closeout", "runtime_ledger_post_window_closeout"):
-        return True
-    for payload in _source_decision_priority_payloads(row):
-        metadata = _as_mapping(payload.get("paper_route_probe_exit"))
-        if _direct_text(metadata, "mode") == "paper_route_exit":
-            return True
-    return False
-
-
-def _source_decision_identifier_values(row: Mapping[str, object]) -> set[str]:
-    return _text_values(
-        row,
-        "trade_decision_id",
-        "decision_id",
-        "decision_hash",
-    )
-
-
-def _paper_route_probe_exit_identifiers(
-    rows: Sequence[Mapping[str, object]],
-) -> set[str]:
-    identifiers: set[str] = set()
-    for row in rows:
-        if _source_row_is_paper_route_probe_exit(row):
-            identifiers.update(_source_decision_identifier_values(row))
-    return identifiers
-
-
-def _source_row_is_paper_route_probe_exit_or_linked(
-    row: Mapping[str, object],
-    *,
-    paper_route_probe_exit_identifiers: set[str],
-) -> bool:
-    if _source_row_is_paper_route_probe_exit(row):
-        return True
-    return bool(
-        _source_decision_identifier_values(row) & paper_route_probe_exit_identifiers
-    )
-
-
-def _source_decision_mode_counts(
-    rows: Sequence[Mapping[str, object]],
-) -> dict[str, int]:
-    counts: dict[str, int] = {}
-    paper_route_probe_exit_identifiers = _paper_route_probe_exit_identifiers(rows)
-    for row in rows:
-        if _source_row_is_paper_route_probe_exit_or_linked(
-            row,
-            paper_route_probe_exit_identifiers=paper_route_probe_exit_identifiers,
-        ):
-            continue
-        mode = _source_decision_mode(row)
-        if mode is None:
-            continue
-        counts[mode] = counts.get(mode, 0) + 1
-    return dict(sorted(counts.items()))
-
-
-def _source_decision_mode_lookup(
-    rows: Sequence[Mapping[str, object]],
-) -> dict[str, str]:
-    modes_by_identifier: dict[str, str] = {}
-    conflicting_identifiers: set[str] = set()
-    for row in rows:
-        mode = _source_decision_mode(row)
-        if mode is None:
-            continue
-        for identifier in _source_decision_identifier_values(row):
-            existing = modes_by_identifier.get(identifier)
-            if existing is not None and existing != mode:
-                conflicting_identifiers.add(identifier)
-                continue
-            modes_by_identifier[identifier] = mode
-    for identifier in conflicting_identifiers:
-        modes_by_identifier.pop(identifier, None)
-    return modes_by_identifier
-
-
-def _source_decision_mode_partition_key(
-    row: Mapping[str, object],
-    *,
-    modes_by_identifier: Mapping[str, str],
-) -> str:
-    explicit = _source_decision_mode(row)
-    if explicit is not None:
-        return explicit
-    inferred_modes = {
-        modes_by_identifier[identifier]
-        for identifier in _source_decision_identifier_values(row)
-        if identifier in modes_by_identifier
-    }
-    if len(inferred_modes) == 1:
-        return next(iter(inferred_modes))
-    return SOURCE_DECISION_MODE_MISSING_PARTITION
-
-
-def _partition_runtime_source_rows_by_decision_mode(
-    *,
-    execution_rows: list[dict[str, object]],
-    decision_lifecycle_rows: list[dict[str, object]] | None,
-    order_lifecycle_rows: list[dict[str, object]] | None,
-    unlinked_order_lifecycle_rows: list[dict[str, object]] | None = None,
-    carry_in_execution_rows: list[dict[str, object]] | None = None,
-    carry_in_decision_lifecycle_rows: list[dict[str, object]] | None = None,
-    carry_in_order_lifecycle_rows: list[dict[str, object]] | None = None,
-) -> (
-    list[
-        tuple[
-            str,
-            list[dict[str, object]],
-            list[dict[str, object]] | None,
-            list[dict[str, object]] | None,
-            list[dict[str, object]] | None,
-            list[dict[str, object]] | None,
-            list[dict[str, object]] | None,
-            list[dict[str, object]] | None,
-        ]
-    ]
-    | None
-):
-    current_source_rows: list[dict[str, object]] = [
-        *execution_rows,
-        *(decision_lifecycle_rows or []),
-        *(order_lifecycle_rows or []),
-        *(unlinked_order_lifecycle_rows or []),
-    ]
-    source_rows: list[dict[str, object]] = [
-        *current_source_rows,
-        *(carry_in_execution_rows or []),
-        *(carry_in_decision_lifecycle_rows or []),
-        *(carry_in_order_lifecycle_rows or []),
-    ]
-    if not source_rows:
-        return None
-    modes_by_identifier = _source_decision_mode_lookup(source_rows)
-    paper_route_probe_exit_identifiers = _paper_route_probe_exit_identifiers(
-        source_rows
-    )
-    partition_relevant_source_rows = [
-        row
-        for row in current_source_rows
-        if not _source_row_is_paper_route_probe_exit_or_linked(
-            row,
-            paper_route_probe_exit_identifiers=paper_route_probe_exit_identifiers,
-        )
-    ]
-    relevant_mode_keys = {
-        _source_decision_mode_partition_key(
-            row,
-            modes_by_identifier=modes_by_identifier,
-        )
-        for row in partition_relevant_source_rows
-    }
-    if len(relevant_mode_keys) <= 1:
-        return None
-    mode_keys = {
-        _source_decision_mode_partition_key(
-            row,
-            modes_by_identifier=modes_by_identifier,
-        )
-        for row in source_rows
-    }
-
-    partitions: list[
-        tuple[
-            str,
-            list[dict[str, object]],
-            list[dict[str, object]] | None,
-            list[dict[str, object]] | None,
-            list[dict[str, object]] | None,
-            list[dict[str, object]] | None,
-            list[dict[str, object]] | None,
-            list[dict[str, object]] | None,
-        ]
-    ] = []
-    for mode_key in sorted(mode_keys):
-        partition_execution_rows = [
-            row
-            for row in execution_rows
-            if _source_decision_mode_partition_key(
-                row,
-                modes_by_identifier=modes_by_identifier,
-            )
-            == mode_key
-        ]
-        partition_decision_rows = [
-            row
-            for row in decision_lifecycle_rows or []
-            if _source_decision_mode_partition_key(
-                row,
-                modes_by_identifier=modes_by_identifier,
-            )
-            == mode_key
-        ]
-        partition_order_rows = [
-            row
-            for row in order_lifecycle_rows or []
-            if _source_decision_mode_partition_key(
-                row,
-                modes_by_identifier=modes_by_identifier,
-            )
-            == mode_key
-        ]
-        partition_unlinked_order_rows = []
-        for row in unlinked_order_lifecycle_rows or []:
-            partition_key = _source_decision_mode_partition_key(
-                row,
-                modes_by_identifier=modes_by_identifier,
-            )
-            if partition_key == mode_key or (
-                partition_key == SOURCE_DECISION_MODE_MISSING_PARTITION
-                and mode_key != SOURCE_DECISION_MODE_MISSING_PARTITION
-            ):
-                partition_unlinked_order_rows.append(row)
-        partition_carry_in_execution_rows = [
-            row
-            for row in carry_in_execution_rows or []
-            if _source_decision_mode_partition_key(
-                row,
-                modes_by_identifier=modes_by_identifier,
-            )
-            == mode_key
-        ]
-        partition_carry_in_decision_rows = [
-            row
-            for row in carry_in_decision_lifecycle_rows or []
-            if _source_decision_mode_partition_key(
-                row,
-                modes_by_identifier=modes_by_identifier,
-            )
-            == mode_key
-        ]
-        partition_carry_in_order_rows = [
-            row
-            for row in carry_in_order_lifecycle_rows or []
-            if _source_decision_mode_partition_key(
-                row,
-                modes_by_identifier=modes_by_identifier,
-            )
-            == mode_key
-        ]
-        partitions.append(
-            (
-                mode_key,
-                partition_execution_rows,
-                partition_decision_rows or None,
-                partition_order_rows or None,
-                partition_unlinked_order_rows or None,
-                partition_carry_in_execution_rows or None,
-                partition_carry_in_decision_rows or None,
-                partition_carry_in_order_rows or None,
-            )
-        )
-    return partitions
-
-
-def _source_decision_rows_profit_proof_eligible(
-    rows: Sequence[Mapping[str, object]],
-) -> bool:
-    modes = _source_decision_mode_counts(rows)
-    if source_decision_mode_counts_have_non_profit_proof_modes(modes):
-        return False
-    explicit: list[bool] = []
-    paper_route_probe_exit_identifiers = _paper_route_probe_exit_identifiers(rows)
-    for row in rows:
-        if _source_row_is_paper_route_probe_exit_or_linked(
-            row,
-            paper_route_probe_exit_identifiers=paper_route_probe_exit_identifiers,
-        ):
-            continue
-        value = _source_decision_profit_proof_flag(row)
-        if value is not None:
-            explicit.append(value)
-    if any(value is False for value in explicit):
-        return False
-    return source_decision_mode_counts_have_profit_proof_modes(modes) or any(
-        value is True for value in explicit
-    )
-
-
-def _source_row_matches_lineage(
-    row: Mapping[str, object],
-    *,
-    candidate_id: str | None,
-    hypothesis_id: str | None,
-    require_source_lineage: bool,
-) -> bool:
-    if not require_source_lineage:
-        return True
-    if candidate_id is not None and candidate_id not in _text_values(
-        row, *SOURCE_LINEAGE_CANDIDATE_KEYS
-    ):
-        return False
-    if hypothesis_id is not None and hypothesis_id not in _text_values(
-        row, *SOURCE_LINEAGE_HYPOTHESIS_KEYS
-    ):
-        return False
-    return True
-
-
-def _source_row_lineage_missing_or_matches(
-    row: Mapping[str, object],
-    *,
-    candidate_id: str | None,
-    hypothesis_id: str | None,
-) -> bool:
-    if candidate_id is not None:
-        candidate_values = _text_values(row, *SOURCE_LINEAGE_CANDIDATE_KEYS)
-        if candidate_values and candidate_id not in candidate_values:
-            return False
-    if hypothesis_id is not None:
-        hypothesis_values = _text_values(row, *SOURCE_LINEAGE_HYPOTHESIS_KEYS)
-        if hypothesis_values and hypothesis_id not in hypothesis_values:
-            return False
-    return True
-
-
-def _source_row_time_before(
-    row: Mapping[str, object],
-    *,
-    window_end: datetime,
-) -> bool:
-    event_time = _runtime_ledger_row_time(row)
-    return event_time is None or event_time < window_end
-
-
-def _source_row_time_after_window(
-    row: Mapping[str, object],
-    *,
-    window_end: datetime,
-    closeout_end: datetime,
-) -> bool:
-    event_time = _runtime_ledger_row_time(row)
-    return (
-        event_time is not None
-        and event_time > window_end
-        and event_time <= closeout_end
-    )
-
-
-def _runtime_open_qtys_by_symbol_from_execution_rows(
-    rows: Sequence[Mapping[str, object]],
-) -> dict[str, Decimal]:
-    open_qtys: dict[str, Decimal] = {}
-    for row in rows:
-        if not _execution_row_has_fill(row):
-            continue
-        symbol = _runtime_source_row_symbol(row)
-        if symbol is None:
-            continue
-        signed_qty = _execution_signed_qty(
-            side=_first_text(row, "side", "action", "order_side"),
-            qty=_first_positive_decimal(
-                row,
-                "filled_qty",
-                "filled_quantity",
-                "qty",
-                "quantity",
-            ),
-        )
-        if signed_qty == 0:
-            continue
-        open_qtys[symbol] = open_qtys.get(symbol, Decimal("0")) + signed_qty
-    return {symbol: qty for symbol, qty in sorted(open_qtys.items()) if qty != 0}
-
-
-def _source_decision_action_offsets_open_qty(
-    row: Mapping[str, object],
-    *,
-    open_qtys: Mapping[str, Decimal],
-) -> bool:
-    symbol = _runtime_source_row_symbol(row)
-    if symbol is None:
-        return False
-    open_qty = open_qtys.get(symbol, Decimal("0"))
-    if open_qty == 0:
-        return False
-    action = (_first_text(row, "action", "side", "order_side") or "").lower()
-    if open_qty > 0:
-        return action in {"sell", "short", "sell_short"}
-    return action in {"buy", "cover", "buy_to_cover"}
-
-
-def _source_row_decision_ids(row: Mapping[str, object]) -> set[str]:
-    return set(
-        _source_identifier_values(
-            [row],
-            "trade_decision_id",
-            "decision_id",
-            "decision_hash",
-        )
-    )
-
-
-def _source_row_execution_ids(row: Mapping[str, object]) -> set[str]:
-    return set(_source_identifier_values([row], "execution_id"))
-
-
-def _mark_post_window_closeout_source_row(
-    row: Mapping[str, object],
-) -> dict[str, object]:
-    return {
-        **dict(row),
-        "post_window_closeout": True,
-        "runtime_ledger_post_window_closeout": True,
-        "runtime_ledger_closeout_reason": "post_window_fill_offsets_window_open_qty",
-    }
-
-
-def _filter_source_rows_for_runtime_window(
-    *,
-    decision_rows: list[dict[str, object]],
-    execution_rows: list[dict[str, object]],
-    order_lifecycle_rows: list[dict[str, object]],
-    window_end: datetime,
-    closeout_end: datetime,
-    candidate_id: str | None,
-    hypothesis_id: str | None,
-    require_source_lineage: bool,
-) -> tuple[
-    list[dict[str, object]],
-    list[dict[str, object]],
-    list[dict[str, object]],
-    dict[str, object],
-]:
-    lineaged_decision_rows = [
-        row
-        for row in decision_rows
-        if _source_row_time_before(row, window_end=window_end)
-        and _source_row_matches_lineage(
-            row,
-            candidate_id=candidate_id,
-            hypothesis_id=hypothesis_id,
-            require_source_lineage=require_source_lineage,
-        )
-    ]
-    lineaged_execution_rows = [
-        row
-        for row in execution_rows
-        if _source_row_time_before(row, window_end=window_end)
-        and _source_row_matches_lineage(
-            row,
-            candidate_id=candidate_id,
-            hypothesis_id=hypothesis_id,
-            require_source_lineage=require_source_lineage,
-        )
-    ]
-    lineaged_order_rows = [
-        row
-        for row in order_lifecycle_rows
-        if _source_row_time_before(row, window_end=window_end)
-        and _source_row_matches_lineage(
-            row,
-            candidate_id=candidate_id,
-            hypothesis_id=hypothesis_id,
-            require_source_lineage=require_source_lineage,
-        )
-    ]
-    diagnostics: dict[str, object] = {
-        "post_window_closeout_lookahead_seconds": int(
-            RUNTIME_LEDGER_POST_WINDOW_CLOSEOUT_LOOKAHEAD.total_seconds()
-        ),
-        "post_window_closeout_query_end": closeout_end.isoformat(),
-    }
-    if closeout_end <= window_end:
-        diagnostics.update(
-            {
-                "post_window_closeout_open_qty_by_symbol": {},
-                "post_window_closeout_decision_count": 0,
-                "post_window_closeout_execution_count": 0,
-                "post_window_closeout_order_event_count": 0,
-            }
-        )
-        return (
-            lineaged_decision_rows,
-            lineaged_execution_rows,
-            lineaged_order_rows,
-            diagnostics,
-        )
-
-    open_qtys = _runtime_open_qtys_by_symbol_from_execution_rows(
-        lineaged_execution_rows
-    )
-    diagnostics["post_window_closeout_open_qty_by_symbol"] = {
-        symbol: str(qty) for symbol, qty in open_qtys.items()
-    }
-    if not open_qtys:
-        diagnostics.update(
-            {
-                "post_window_closeout_decision_count": 0,
-                "post_window_closeout_execution_count": 0,
-                "post_window_closeout_order_event_count": 0,
-            }
-        )
-        return (
-            lineaged_decision_rows,
-            lineaged_execution_rows,
-            lineaged_order_rows,
-            diagnostics,
-        )
-
-    closeout_decision_rows = [
-        row
-        for row in decision_rows
-        if _source_row_time_after_window(
-            row,
-            window_end=window_end,
-            closeout_end=closeout_end,
-        )
-        and source_decision_mode_is_profit_proof_eligible(_source_decision_mode(row))
-        and _source_row_lineage_missing_or_matches(
-            row,
-            candidate_id=candidate_id,
-            hypothesis_id=hypothesis_id,
-        )
-        and _source_decision_action_offsets_open_qty(row, open_qtys=open_qtys)
-    ]
-    closeout_decision_ids = set[str]()
-    for row in closeout_decision_rows:
-        closeout_decision_ids.update(_source_row_decision_ids(row))
-    if not closeout_decision_ids:
-        diagnostics.update(
-            {
-                "post_window_closeout_decision_count": 0,
-                "post_window_closeout_execution_count": 0,
-                "post_window_closeout_order_event_count": 0,
-            }
-        )
-        return (
-            lineaged_decision_rows,
-            lineaged_execution_rows,
-            lineaged_order_rows,
-            diagnostics,
-        )
-
-    closeout_execution_rows = [
-        row
-        for row in execution_rows
-        if _source_row_time_after_window(
-            row,
-            window_end=window_end,
-            closeout_end=closeout_end,
-        )
-        and _source_row_decision_ids(row) & closeout_decision_ids
-        and _execution_row_has_fill(row)
-        and _execution_query_row_has_tca_ref(row)
-    ]
-    closeout_execution_decision_ids = set[str]()
-    closeout_execution_ids = set[str]()
-    for row in closeout_execution_rows:
-        closeout_execution_decision_ids.update(_source_row_decision_ids(row))
-        closeout_execution_ids.update(_source_row_execution_ids(row))
-
-    closeout_order_rows = [
-        row
-        for row in order_lifecycle_rows
-        if _source_row_time_after_window(
-            row,
-            window_end=window_end,
-            closeout_end=closeout_end,
-        )
-        and (
-            _source_row_decision_ids(row) & closeout_decision_ids
-            or _source_row_execution_ids(row) & closeout_execution_ids
-        )
-        and _source_authority_order_event_row(row)
-    ]
-    fill_order_decision_ids = set[str]()
-    for row in closeout_order_rows:
-        if _runtime_ledger_event_type(row) not in _RUNTIME_LEDGER_FILL_EVENTS:
-            continue
-        fill_order_decision_ids.update(_source_row_decision_ids(row))
-        if _source_row_execution_ids(row) & closeout_execution_ids:
-            fill_order_decision_ids.update(closeout_execution_decision_ids)
-
-    eligible_decision_ids = (
-        closeout_decision_ids
-        & closeout_execution_decision_ids
-        & fill_order_decision_ids
-    )
-    eligible_execution_ids = {
-        execution_id
-        for row in closeout_execution_rows
-        if _source_row_decision_ids(row) & eligible_decision_ids
-        for execution_id in _source_row_execution_ids(row)
-    }
-    eligible_closeout_decision_rows = [
-        _mark_post_window_closeout_source_row(row)
-        for row in closeout_decision_rows
-        if _source_row_decision_ids(row) & eligible_decision_ids
-    ]
-    eligible_closeout_execution_rows = [
-        _mark_post_window_closeout_source_row(row)
-        for row in closeout_execution_rows
-        if _source_row_decision_ids(row) & eligible_decision_ids
-    ]
-    eligible_closeout_order_rows = [
-        _mark_post_window_closeout_source_row(row)
-        for row in closeout_order_rows
-        if (
-            _source_row_decision_ids(row) & eligible_decision_ids
-            or _source_row_execution_ids(row) & eligible_execution_ids
-        )
-    ]
-    diagnostics.update(
-        {
-            "post_window_closeout_decision_count": len(eligible_closeout_decision_rows),
-            "post_window_closeout_execution_count": len(
-                eligible_closeout_execution_rows
-            ),
-            "post_window_closeout_order_event_count": len(eligible_closeout_order_rows),
-        }
-    )
-    return (
-        [*lineaged_decision_rows, *eligible_closeout_decision_rows],
-        [*lineaged_execution_rows, *eligible_closeout_execution_rows],
-        [*lineaged_order_rows, *eligible_closeout_order_rows],
-        diagnostics,
-    )
-
-
-def _runtime_ledger_equity_denominator_from_rows(
-    rows: Sequence[Mapping[str, object]],
-) -> tuple[Decimal, str] | None:
-    denominators: list[tuple[Decimal, str]] = []
-    for row in rows:
-        for key in _RUNTIME_LEDGER_EQUITY_DENOMINATOR_KEYS:
-            value = _first_decimal(row, key)
-            if value is not None and value > 0:
-                denominators.append((value, key))
-                break
-    if not denominators:
-        return None
-    return min(denominators, key=lambda item: item[0])
-
-
-def _json_default(value: object) -> str:
-    if isinstance(value, datetime):
-        return value.isoformat()
-    if isinstance(value, Decimal):
-        return str(value)
-    return str(value)
-
-
-def _stable_payload_digest(value: object) -> str | None:
-    if not isinstance(value, Mapping):
-        return None
-    payload = {str(item_key): item for item_key, item in value.items()}
-    if not payload:
-        return None
-    encoded = json.dumps(
-        payload,
-        sort_keys=True,
-        separators=(",", ":"),
-        default=_json_default,
-    ).encode("utf-8")
-    return hashlib.sha256(encoded).hexdigest()
-
-
-def _runtime_source_lineage_hash(
-    *,
-    candidate_id: str | None,
-    hypothesis_id: str | None,
-) -> str | None:
-    payload = {
-        key: value
-        for key, value in {
-            "candidate_id": candidate_id,
-            "hypothesis_id": hypothesis_id,
-            "source": "runtime_window_lineage_filter",
-        }.items()
-        if value
-    }
-    return _stable_payload_digest(payload)
-
-
-def _attach_source_lineage_context(
-    rows: Sequence[Mapping[str, object]],
-    *,
-    candidate_id: str | None,
-    hypothesis_id: str | None,
-) -> list[dict[str, object]]:
-    lineage_hash = _runtime_source_lineage_hash(
-        candidate_id=candidate_id,
-        hypothesis_id=hypothesis_id,
-    )
-    contextualized: list[dict[str, object]] = []
-    for row in rows:
-        payload = dict(row)
-        if candidate_id:
-            payload.setdefault("source_candidate_id", candidate_id)
-        if hypothesis_id:
-            payload.setdefault("source_hypothesis_id", hypothesis_id)
-        if lineage_hash:
-            payload["lineage_hash"] = lineage_hash
-        contextualized.append(payload)
-    return contextualized
-
-
-def _decimal_ceil_cent(value: Decimal) -> Decimal:
-    if value <= 0:
-        return Decimal("0")
-    return value.quantize(_CENT, rounding=ROUND_CEILING)
-
-
-def _row_has_alpaca_us_equity_order_source(row: Mapping[str, object]) -> bool:
-    alpaca_markers = {
-        str(item or "").strip().lower()
-        for item in (
-            *_text_values(row, "feed", "source_topic", "channel", "submit_path"),
-            *_text_values(row, "source"),
-        )
-    }
-    if not any(
-        "alpaca" in marker or marker == "trade_updates" for marker in alpaca_markers
-    ):
-        return False
-    asset_classes = {
-        value.strip().lower().replace("-", "_")
-        for value in _text_values(row, "asset_class")
-    }
-    return not asset_classes or any(
-        item in {"us_equity", "us_equities", "equity", "equities"}
-        for item in asset_classes
-    )
-
-
-def _alpaca_2026_equity_fee_schedule_cost(
-    row: Mapping[str, object],
-    *,
-    side: Any,
-    filled_qty: Decimal | None,
-    filled_notional: Decimal,
-) -> tuple[Decimal, str] | None:
-    if (
-        filled_qty is None
-        or filled_qty <= 0
-        or filled_notional <= 0
-        or not _row_has_alpaca_us_equity_order_source(row)
-    ):
-        return None
-    normalized_side = str(side or "").strip().lower().replace("-", "_")
-    if normalized_side in {"buy", "buy_to_cover", "cover"}:
-        return Decimal("0"), "alpaca_2026_equity_zero_commission_and_cat_fee_schedule"
-    if normalized_side not in {"sell", "sell_short", "short"}:
-        return None
-    sec_fee = _decimal_ceil_cent(filled_notional * ALPACA_2026_EQUITY_SEC_FEE_RATE)
-    taf_fee = min(
-        _decimal_ceil_cent(filled_qty * ALPACA_2026_EQUITY_TAF_RATE_PER_SHARE),
-        ALPACA_2026_EQUITY_TAF_CAP,
-    )
-    return (
-        sec_fee + taf_fee,
-        "alpaca_2026_equity_sec_taf_cat_fee_schedule",
-    )
-
-
-def _alpaca_2026_equity_fee_schedule_hash() -> str:
-    digest = _stable_payload_digest(
-        {
-            "broker": "alpaca",
-            "asset_class": "us_equity",
-            "schedule": "alpaca_brokerage_fee_schedule",
-            "revised_on": ALPACA_2026_FEE_SCHEDULE_REVISED_ON,
-            "sec_fee_rate_per_notional": str(ALPACA_2026_EQUITY_SEC_FEE_RATE),
-            "taf_rate_per_share": str(ALPACA_2026_EQUITY_TAF_RATE_PER_SHARE),
-            "taf_cap": str(ALPACA_2026_EQUITY_TAF_CAP),
-            "cat_fee_equities": "0",
-            "commission": "0",
-        }
-    )
-    assert digest is not None
-    return digest
-
-
-def _cost_basis_is_alpaca_fee_schedule(value: object) -> bool:
-    return str(value or "").strip().startswith("alpaca_2026_equity_")
-
-
-def _first_payload_digest(row: Mapping[str, object], *keys: str) -> str | None:
-    for payload in _row_payloads(row):
-        for key in keys:
-            if digest := _stable_payload_digest(payload.get(key)):
-                return digest
-    return None
-
-
-_LINEAGE_CONTEXT_KEYS = (
-    "strategy_id",
-    "primary_declared_strategy_id",
-    "source_declared_strategy_ids",
-    "paper_route_target_plan_source",
-    "simulation_run_id",
-    "dataset_id",
-    "dataset_snapshot_ref",
-    "dataset_snapshot_hash",
-    "replay_data_hash",
-    "replay_tape_content_sha256",
-    "source_query_digest",
-    "source_topic",
-    "source_partition",
-)
-
-
-def _first_lineage_digest(row: Mapping[str, object]) -> str | None:
-    if digest := _first_payload_digest(
-        row, "lineage", "candidate_lineage", "source_lineage"
-    ):
-        return digest
-    for payload in _row_payloads(row):
-        lineage_payload = {
-            key: value
-            for key in _LINEAGE_CONTEXT_KEYS
-            if (value := payload.get(key)) is not None
-        }
-        if digest := _stable_payload_digest(lineage_payload):
-            return digest
-    return None
-
-
-def _runtime_ledger_bucket_payload(bucket: RuntimeLedgerBucket) -> dict[str, object]:
-    return {
-        "bucket_started_at": bucket.bucket_started_at.isoformat(),
-        "bucket_ended_at": bucket.bucket_ended_at.isoformat(),
-        "account_label": bucket.account_label,
-        "strategy_id": bucket.strategy_id,
-        "symbol": bucket.symbol,
-        "fill_count": bucket.fill_count,
-        "decision_count": bucket.decision_count,
-        "submitted_order_count": bucket.submitted_order_count,
-        "cancelled_order_count": bucket.cancelled_order_count,
-        "rejected_order_count": bucket.rejected_order_count,
-        "unfilled_order_count": bucket.unfilled_order_count,
-        "closed_trade_count": bucket.closed_trade_count,
-        "open_position_count": bucket.open_position_count,
-        "filled_notional": str(bucket.filled_notional),
-        "gross_strategy_pnl": str(bucket.gross_strategy_pnl),
-        "cost_amount": str(bucket.cost_amount),
-        "net_strategy_pnl_after_costs": str(bucket.net_strategy_pnl_after_costs),
-        "post_cost_expectancy_bps": (
-            str(bucket.post_cost_expectancy_bps)
-            if bucket.post_cost_expectancy_bps is not None
-            else None
-        ),
-        "diagnostic_closed_trade_expectancy_bps": (
-            str(bucket.diagnostic_closed_trade_expectancy_bps)
-            if bucket.diagnostic_closed_trade_expectancy_bps is not None
-            else None
-        ),
-        "diagnostic_closed_trade_expectancy_basis": (
-            "realized_closed_trips_after_explicit_costs_not_promotion_grade"
-            if bucket.diagnostic_closed_trade_expectancy_bps is not None
-            else None
-        ),
-        "cost_basis_counts": bucket.cost_basis_counts,
-        "execution_policy_hash_counts": bucket.execution_policy_hash_counts,
-        "cost_model_hash_counts": bucket.cost_model_hash_counts,
-        "lineage_hash_counts": bucket.lineage_hash_counts,
-        "blockers": bucket.blockers,
-        "ledger_schema_version": bucket.ledger_schema_version,
-        "pnl_basis": bucket.pnl_basis,
-    }
-
-
-def _mapping_hash_count(value: object) -> int:
-    if not isinstance(value, Mapping):
-        return 0
-    return sum(1 for key in value.keys() if str(key).strip())
-
-
-def _positive_count_mapping_present(value: object) -> bool:
-    if not isinstance(value, Mapping):
-        return False
-    for item in cast(Mapping[object, object], value).values():
-        try:
-            parsed = Decimal(str(item))
-        except Exception:
-            continue
-        if parsed.is_finite() and parsed > 0:
-            return True
-    return False
-
-
-def _runtime_ledger_explicit_costs_present(bucket: Mapping[str, object]) -> bool:
-    cost_amount = _decimal_or_none(
-        bucket.get("cost_amount")
-        if bucket.get("cost_amount") is not None
-        else bucket.get("runtime_ledger_cost_amount")
-    )
-    if cost_amount is None or cost_amount < 0:
-        return False
-    if is_non_promotion_grade_runtime_cost_basis(bucket.get("cost_basis")):
-        return False
-    if cost_basis_counts_have_non_promotion_grade_costs(
-        bucket.get("cost_basis_counts")
-    ) or cost_basis_counts_have_non_promotion_grade_costs(
-        bucket.get("post_cost_basis_counts")
-    ):
-        return False
-    if _positive_count_mapping_present(bucket.get("cost_basis_counts")):
-        return True
-    if _positive_count_mapping_present(bucket.get("post_cost_basis_counts")):
-        return True
-    return (
-        _text_or_none(
-            bucket.get("cost_basis")
-            or bucket.get("cost_source")
-            or bucket.get("fee_basis")
-            or bucket.get("commission_basis")
-            or bucket.get("broker_fee_basis")
-        )
-        is not None
-    )
-
-
-def _runtime_ledger_bucket_profit_proof_blockers(
-    bucket: Mapping[str, object],
-) -> list[str]:
-    blockers: list[str] = []
-
-    def add(blocker: str) -> None:
-        if blocker not in blockers:
-            blockers.append(blocker)
-
-    filled_notional = _decimal_or_none(bucket.get("filled_notional"))
-    cost_amount = _decimal_or_none(bucket.get("cost_amount"))
-    post_cost_expectancy = _decimal_or_none(bucket.get("post_cost_expectancy_bps"))
-    diagnostic_closed_trade_expectancy = _decimal_or_none(
-        bucket.get("diagnostic_closed_trade_expectancy_bps")
-    )
-    if bucket.get("pnl_basis") != POST_COST_BASIS_RUNTIME_LEDGER:
-        add("runtime_ledger_pnl_basis_not_runtime_ledger")
-    if bucket.get("ledger_schema_version") not in RUNTIME_LEDGER_BUCKET_SCHEMAS:
-        add("runtime_ledger_schema_not_supported")
-    for blocker in runtime_ledger_promotion_source_authority_blockers(bucket):
-        add(blocker)
-    execution_count = _nonnegative_int(
-        _as_mapping(bucket.get("source_row_counts")).get("executions")
-    )
-    tca_count = _nonnegative_int(
-        _as_mapping(bucket.get("source_row_counts")).get("execution_tca_metrics")
-    )
-    execution_tca_ref_count = len(_runtime_ledger_execution_tca_metric_refs(bucket))
-    execution_tca_required = (
-        _first_bool(bucket, "execution_tca_required", "requires_execution_tca") is True
-        or tca_count > 0
-    )
-    if (
-        execution_tca_required
-        and execution_count > 0
-        and (tca_count < execution_count or execution_tca_ref_count < execution_count)
-    ):
-        add(EXECUTION_TCA_MISSING_BLOCKER)
-        add(RUNTIME_LEDGER_EXECUTION_TCA_REFS_MISSING_BLOCKER)
-    for blocker in _metadata_text_list(bucket.get("blockers")):
-        add(blocker)
-    if _first_bool(bucket, "order_feed_lifecycle_complete") is False:
-        add(ORDER_FEED_LIFECYCLE_MISSING_BLOCKER)
-    if _first_bool(bucket, "execution_economics_complete") is False:
-        add(EXECUTION_ECONOMICS_MISSING_BLOCKER)
-    if _nonnegative_int(bucket.get("fill_count")) <= 0:
-        add("runtime_fills_missing")
-    if _nonnegative_int(bucket.get("decision_count")) <= 0:
-        add("runtime_decision_lifecycle_missing")
-    if _nonnegative_int(bucket.get("submitted_order_count")) <= 0:
-        add("submitted_order_lifecycle_missing")
-    if _nonnegative_int(bucket.get("closed_trade_count")) <= 0:
-        add("closed_round_trip_missing")
-    if bucket.get("open_position_count") is None:
-        add("runtime_ledger_open_position_count_missing")
-    if _nonnegative_int(bucket.get("open_position_count")) > 0:
-        add("unclosed_position")
-    if filled_notional is None or filled_notional <= 0:
-        add("filled_notional_missing")
-    if not _runtime_ledger_explicit_costs_present(bucket):
-        add("runtime_ledger_explicit_costs_missing")
-    if cost_amount is None or cost_amount < 0:
-        add(
-            "runtime_ledger_cost_amount_missing"
-            if cost_amount is None
-            else "runtime_ledger_cost_amount_negative"
-        )
-    if is_non_promotion_grade_runtime_cost_basis(bucket.get("cost_basis")):
-        add("runtime_ledger_cost_basis_non_promotion_grade")
-    if cost_basis_counts_have_non_promotion_grade_costs(
-        bucket.get("cost_basis_counts")
-    ):
-        add("runtime_ledger_cost_basis_non_promotion_grade")
-    source_decision_mode = normalize_source_decision_mode(
-        bucket.get("source_decision_mode")
-    )
-    profit_proof_eligible = _first_bool(bucket, "profit_proof_eligible")
-    source_decision_mode_counts = bucket.get("source_decision_mode_counts")
-    if (
-        source_decision_mode is None
-        and not source_decision_mode_counts_have_profit_proof_modes(
-            source_decision_mode_counts
-        )
-        and not source_decision_mode_counts_have_non_profit_proof_modes(
-            source_decision_mode_counts
-        )
-        and profit_proof_eligible is not True
-    ):
-        add(SOURCE_DECISION_MODE_PROFIT_PROOF_MISSING_BLOCKER)
-    if source_decision_mode and not source_decision_mode_is_profit_proof_eligible(
-        source_decision_mode
-    ):
-        add(SOURCE_DECISION_MODE_NOT_PROFIT_PROOF_ELIGIBLE_BLOCKER)
-    if source_decision_mode_counts_have_non_profit_proof_modes(
-        source_decision_mode_counts
-    ):
-        add(SOURCE_DECISION_MODE_NOT_PROFIT_PROOF_ELIGIBLE_BLOCKER)
-    if profit_proof_eligible is False:
-        add(SOURCE_DECISION_MODE_NOT_PROFIT_PROOF_ELIGIBLE_BLOCKER)
-    if post_cost_expectancy is None:
-        add(
-            "runtime_ledger_post_cost_expectancy_not_promotion_grade"
-            if diagnostic_closed_trade_expectancy is not None
-            else "runtime_ledger_post_cost_expectancy_missing"
-        )
-    if _mapping_hash_count(bucket.get("execution_policy_hash_counts")) <= 0:
-        add("runtime_ledger_execution_policy_hash_missing")
-    if _mapping_hash_count(bucket.get("cost_model_hash_counts")) <= 0:
-        add("runtime_ledger_cost_model_hash_missing")
-    if _mapping_hash_count(bucket.get("lineage_hash_counts")) <= 0:
-        add("runtime_ledger_lineage_hash_missing")
-    return blockers
-
-
-def _runtime_ledger_bucket_profit_proof_present(
-    bucket: Mapping[str, object],
-) -> bool:
-    return not _runtime_ledger_bucket_profit_proof_blockers(bucket)
-
-
-def _runtime_ledger_tca_row_from_bucket(
-    *,
-    bucket: RuntimeLedgerBucket,
-    computed_at: datetime | None = None,
-) -> dict[str, object]:
-    payload = _runtime_ledger_bucket_payload(bucket)
-    promotion_eligible = _runtime_ledger_bucket_profit_proof_present(payload)
-    return {
-        "computed_at": computed_at
-        or max(
-            bucket.bucket_started_at,
-            bucket.bucket_ended_at - timedelta(microseconds=1),
-        ),
-        "abs_slippage_bps": None,
-        "explicit_cost_bps": (
-            (bucket.cost_amount / bucket.filled_notional) * Decimal("10000")
-            if bucket.filled_notional > 0
-            else None
-        ),
-        "post_cost_expectancy_bps": bucket.post_cost_expectancy_bps,
-        "post_cost_expectancy_basis": POST_COST_BASIS_RUNTIME_LEDGER,
-        "post_cost_promotion_eligible": promotion_eligible,
-        "diagnostic_closed_trade_expectancy_bps": (
-            bucket.diagnostic_closed_trade_expectancy_bps
-        ),
-        "diagnostic_closed_trade_expectancy_basis": (
-            "realized_closed_trips_after_explicit_costs_not_promotion_grade"
-            if bucket.diagnostic_closed_trade_expectancy_bps is not None
-            else None
-        ),
-        "realized_gross_pnl": bucket.gross_strategy_pnl,
-        "realized_net_pnl": bucket.net_strategy_pnl_after_costs,
-        "turnover_notional": bucket.filled_notional,
-        "runtime_ledger_blockers": bucket.blockers,
-        "runtime_ledger_cost_basis_counts": bucket.cost_basis_counts,
-        "runtime_ledger_pnl_basis": bucket.pnl_basis,
-        "runtime_ledger_bucket": payload,
-    }
-
-
-def _append_runtime_ledger_tca_row_blocker(
-    *,
-    tca_rows: list[dict[str, object]],
-    blocker: str,
-) -> None:
-    for row in tca_rows:
-        bucket = row.get("runtime_ledger_bucket")
-        if not isinstance(bucket, Mapping):
-            continue
-        bucket_payload = {str(key): item for key, item in bucket.items()}
-        blockers = _metadata_text_list(bucket_payload.get("blockers"))
-        if blocker not in blockers:
-            blockers.append(blocker)
-        bucket_payload["blockers"] = blockers
-        row["runtime_ledger_bucket"] = bucket_payload
-        row["runtime_ledger_blockers"] = blockers
-        row["post_cost_promotion_eligible"] = False
-
-
-def _with_runtime_ledger_source_authority_context(
-    bucket: Mapping[str, object],
-    *,
-    source_window_start: datetime,
-    source_window_end: datetime,
-    source_refs: Sequence[str],
-    source_row_counts: Mapping[str, int],
-    trade_decision_ids: Sequence[object] = (),
-    execution_ids: Sequence[object] = (),
-    execution_tca_metric_ids: Sequence[object] = (),
-    execution_order_event_ids: Sequence[object] = (),
-    source_window_ids: Sequence[object] = (),
-    source_offsets: Sequence[Mapping[str, object]] = (),
-    source_window_status_counts: Mapping[str, int] | None = None,
-    source_window_classification_counts: Mapping[str, int] | None = None,
-    source_window_gap_count: int = 0,
-    source_window_gap_ranges: Sequence[Mapping[str, object]] = (),
-    order_feed_lifecycle_complete: bool | None = None,
-    execution_economics_complete: bool | None = None,
-    execution_tca_required: bool | None = None,
-    source_materialization: str | None = None,
-    authority_class: str | None = None,
-    authority_reason: str | None = None,
-) -> dict[str, object]:
-    payload = dict(bucket)
-    payload.setdefault("source_window_start", source_window_start.isoformat())
-    payload.setdefault("source_window_end", source_window_end.isoformat())
-    if source_refs:
-        existing_refs = _metadata_text_list(payload.get("source_refs"))
-        payload["source_refs"] = list(dict.fromkeys([*existing_refs, *source_refs]))
-    for key, values in (
-        ("trade_decision_ids", trade_decision_ids),
-        ("execution_ids", execution_ids),
-        ("execution_tca_metric_ids", execution_tca_metric_ids),
-        ("execution_order_event_ids", execution_order_event_ids),
-        ("source_window_ids", source_window_ids),
-    ):
-        merged_values = _metadata_text_list(payload.get(key))
-        merged_values.extend(
-            _text for value in values if (_text := _text_or_none(value))
-        )
-        if merged_values:
-            payload[key] = list(dict.fromkeys(merged_values))
-    for source_key, alias_keys in (
-        (
-            "trade_decision_ids",
-            ("trade_decision_refs", "decision_ids", "decision_refs"),
-        ),
-        ("execution_ids", ("execution_refs", "runtime_ledger_execution_ids")),
-        (
-            "execution_tca_metric_ids",
-            (
-                "execution_tca_metric_refs",
-                "runtime_ledger_execution_tca_metric_ids",
-                "runtime_ledger_execution_tca_metric_refs",
-            ),
-        ),
-        (
-            "execution_order_event_ids",
-            (
-                "execution_order_event_refs",
-                "runtime_ledger_execution_order_event_ids",
-                "runtime_ledger_execution_order_event_refs",
-            ),
-        ),
-        (
-            "source_window_ids",
-            (
-                "source_window_refs",
-                "runtime_ledger_source_window_ids",
-                "runtime_ledger_source_window_refs",
-            ),
-        ),
-    ):
-        values = _metadata_text_list(payload.get(source_key))
-        if not values:
-            continue
-        for alias_key in alias_keys:
-            existing_values = _metadata_text_list(payload.get(alias_key))
-            payload[alias_key] = list(dict.fromkeys([*existing_values, *values]))
-    if source_offsets:
-        existing_offsets = [
-            dict(item)
-            for item in cast(Sequence[object], payload.get("source_offsets") or [])
-            if isinstance(item, Mapping)
-        ]
-        seen_offsets = {
-            (
-                str(item.get("topic") or ""),
-                str(item.get("partition") or ""),
-                str(item.get("offset") or ""),
-            )
-            for item in existing_offsets
-        }
-        for item in source_offsets:
-            offset = {
-                "topic": item.get("topic"),
-                "partition": item.get("partition"),
-                "offset": item.get("offset"),
-            }
-            topic = offset.get("topic")
-            partition = offset.get("partition")
-            source_offset = offset.get("offset")
-            key = (
-                str(topic or ""),
-                "" if partition is None else str(partition),
-                "" if source_offset is None else str(source_offset),
-            )
-            if all(key) and key not in seen_offsets:
-                existing_offsets.append(offset)
-                seen_offsets.add(key)
-        if existing_offsets:
-            payload["source_offsets"] = existing_offsets
-    if source_window_status_counts:
-        payload["source_window_status_counts"] = _merge_count_mappings(
-            _as_mapping(payload.get("source_window_status_counts")),
-            source_window_status_counts,
-        )
-    if source_window_classification_counts:
-        payload["source_window_classification_counts"] = _merge_count_mappings(
-            _as_mapping(payload.get("source_window_classification_counts")),
-            source_window_classification_counts,
-        )
-    if source_window_gap_count > 0:
-        payload["source_window_gap_count"] = max(
-            _nonnegative_int(payload.get("source_window_gap_count")),
-            source_window_gap_count,
-        )
-    if source_window_gap_ranges:
-        existing_gap_ranges = [
-            dict(item)
-            for item in cast(
-                Sequence[object], payload.get("source_window_gap_ranges") or []
-            )
-            if isinstance(item, Mapping)
-        ]
-        existing_gap_ranges.extend(dict(item) for item in source_window_gap_ranges)
-        if existing_gap_ranges:
-            payload["source_window_gap_ranges"] = existing_gap_ranges
-    if order_feed_lifecycle_complete is not None:
-        payload["order_feed_lifecycle_complete"] = order_feed_lifecycle_complete
-    if execution_economics_complete is not None:
-        payload["execution_economics_complete"] = execution_economics_complete
-    if execution_tca_required is not None:
-        payload["execution_tca_required"] = execution_tca_required
-    if source_materialization:
-        payload["source_materialization"] = source_materialization
-    if authority_class:
-        payload["authority_class"] = authority_class
-    if authority_reason:
-        payload["authority_reason"] = authority_reason
-    existing_counts = _as_mapping(payload.get("source_row_counts"))
-    merged_counts: dict[str, int] = {
-        str(key): _nonnegative_int(value) for key, value in existing_counts.items()
-    }
-    for key, value in source_row_counts.items():
-        table_name = str(key)
-        merged_counts[table_name] = max(
-            merged_counts.get(table_name, 0),
-            max(0, int(value)),
-        )
-    if merged_counts:
-        payload["source_row_counts"] = dict(sorted(merged_counts.items()))
-    existing_blockers = _metadata_text_list(payload.get("blockers"))
-    if existing_blockers:
-        refreshed_source_blockers = set(
-            runtime_ledger_promotion_source_authority_blockers(payload)
-        )
-        refreshed_blockers = [
-            blocker
-            for blocker in existing_blockers
-            if (
-                blocker not in _RUNTIME_LEDGER_SOURCE_AUTHORITY_BLOCKERS
-                or blocker in refreshed_source_blockers
-            )
-        ]
-        for blocker in refreshed_source_blockers:
-            if blocker not in refreshed_blockers:
-                refreshed_blockers.append(blocker)
-        payload["blockers"] = refreshed_blockers
-    return payload
-
-
-def _merge_count_mappings(
-    existing: Mapping[object, object],
-    incoming: Mapping[str, int],
-) -> dict[str, int]:
-    merged = {str(key): _nonnegative_int(value) for key, value in existing.items()}
-    for key, value in incoming.items():
-        merged[str(key)] = max(0, int(value)) + merged.get(str(key), 0)
-    return dict(sorted((key, value) for key, value in merged.items() if value > 0))
-
-
-def _source_identifier_values(
-    rows: Sequence[Mapping[str, object]] | None,
-    *keys: str,
-) -> list[str]:
-    values: list[str] = []
-    for row in rows or ():
-        for key in keys:
-            raw_value = row.get(key)
-            row_values: list[str] = []
-            if isinstance(raw_value, Sequence) and not isinstance(
-                raw_value,
-                (str, bytes, bytearray),
-            ):
-                row_values = [
-                    value
-                    for item in raw_value
-                    if (value := _text_or_none(item)) is not None
-                ]
-            elif (value := _text_or_none(raw_value)) is not None:
-                row_values = [value]
-            if row_values:
-                values.extend(row_values)
-                break
-    return list(dict.fromkeys(values))
-
-
-def _source_offset_values(
-    rows: Sequence[Mapping[str, object]] | None,
-) -> list[dict[str, object]]:
-    offsets: list[dict[str, object]] = []
-    seen: set[tuple[str, str, str]] = set()
-    for row in rows or ():
-        topic = _text_or_none(row.get("source_topic"))
-        partition = row.get("source_partition")
-        offset = row.get("source_offset")
-        if topic is None or partition is None or offset is None:
-            continue
-        key = (topic, str(partition), str(offset))
-        if key in seen:
-            continue
-        offsets.append({"topic": topic, "partition": partition, "offset": offset})
-        seen.add(key)
-    return offsets
-
-
-def _with_canonical_runtime_source_refs(
-    row: Mapping[str, object],
-) -> dict[str, object]:
-    normalized = {str(key): value for key, value in row.items()}
-    if normalized.get("execution_order_event_id") is None:
-        execution_order_event_ids = _source_identifier_values(
-            [normalized],
-            *_EXECUTION_ORDER_EVENT_REF_KEYS,
-        )
-        if execution_order_event_ids:
-            normalized["execution_order_event_id"] = execution_order_event_ids[0]
-    if normalized.get("source_window_id") is None:
-        source_window_ids = _source_identifier_values(
-            [normalized],
-            *_SOURCE_WINDOW_REF_KEYS,
-        )
-        if source_window_ids:
-            normalized["source_window_id"] = source_window_ids[0]
-    return normalized
-
-
-def _unique_source_window_rows(
-    rows: Sequence[Mapping[str, object]] | None,
-) -> list[Mapping[str, object]]:
-    unique_rows: list[Mapping[str, object]] = []
-    seen: set[str] = set()
-    for row in rows or ():
-        source_window_id = _text_or_none(row.get("source_window_id"))
-        if source_window_id is None:
-            continue
-        if source_window_id in seen:
-            continue
-        seen.add(source_window_id)
-        unique_rows.append(row)
-    return unique_rows
-
-
-def _source_window_status_counts(
-    rows: Sequence[Mapping[str, object]] | None,
-) -> dict[str, int]:
-    counts: dict[str, int] = {}
-    for row in _unique_source_window_rows(rows):
-        status = _text_or_none(row.get("source_window_status"))
-        if status is None:
-            continue
-        counts[status] = counts.get(status, 0) + 1
-    return dict(sorted(counts.items()))
-
-
-def _source_window_classification_counts(
-    rows: Sequence[Mapping[str, object]] | None,
-) -> dict[str, int]:
-    field_names = {
-        "inserted": "source_window_inserted_count",
-        "duplicate": "source_window_duplicate_count",
-        "malformed_json": "source_window_malformed_count",
-        "missing_trade_update_payload": "source_window_missing_payload_count",
-        "missing_order_identity": "source_window_missing_identity_count",
-        "out_of_scope_account": "source_window_out_of_scope_account_count",
-        "unlinked_execution": "source_window_unlinked_execution_count",
-        "unlinked_decision": "source_window_unlinked_decision_count",
-        "failed_unhandled": "source_window_failed_unhandled_count",
-        "dropped": "source_window_dropped_count",
-    }
-    counts: dict[str, int] = {}
-    for row in _unique_source_window_rows(rows):
-        for classification, field_name in field_names.items():
-            count = _nonnegative_int(row.get(field_name))
-            if count > 0:
-                counts[classification] = counts.get(classification, 0) + count
-    return dict(sorted(counts.items()))
-
-
-def _source_window_gap_count(
-    rows: Sequence[Mapping[str, object]] | None,
-) -> int:
-    return sum(
-        _nonnegative_int(row.get("source_window_gap_count"))
-        for row in _unique_source_window_rows(rows)
-    )
-
-
-def _source_window_gap_ranges(
-    rows: Sequence[Mapping[str, object]] | None,
-) -> list[dict[str, object]]:
-    ranges: list[dict[str, object]] = []
-    for row in _unique_source_window_rows(rows):
-        gap_ranges = row.get("source_window_gap_ranges")
-        if not isinstance(gap_ranges, Sequence) or isinstance(
-            gap_ranges, (str, bytes, bytearray)
-        ):
-            continue
-        for item in cast(Sequence[object], gap_ranges):
-            if isinstance(item, Mapping):
-                ranges.append(dict(item))
-    return ranges
-
-
-def _decision_lifecycle_query_row(row: Sequence[object]) -> dict[str, object]:
-    if len(row) >= 7:
-        return {
-            "trade_decision_id": str(row[0]),
-            "computed_at": row[1],
-            "event_type": "decision",
-            "symbol": row[2],
-            "account_label": row[3],
-            "strategy_id": row[4],
-            "decision_hash": row[5],
-            "decision_json": row[6],
-        }
-    return {
-        "trade_decision_id": str(row[4]),
-        "computed_at": row[0],
-        "event_type": "decision",
-        "symbol": row[1],
-        "account_label": row[2],
-        "strategy_id": row[3],
-        "decision_hash": row[4],
-        "decision_json": row[5],
-    }
-
-
-def _decision_lifecycle_query_row_has_time(row: Sequence[object]) -> bool:
-    return (row[1] if len(row) >= 7 else row[0]) is not None
-
-
-def _execution_query_row(row: Sequence[object]) -> dict[str, object]:
-    if len(row) >= 20:
-        return {
-            "execution_id": str(row[0]),
-            "trade_decision_id": str(row[1]),
-            "decision_created_at": row[2],
-            "computed_at": row[3],
-            "execution_event_at": row[3],
-            "execution_created_at": row[4],
-            "symbol": row[5],
-            "side": row[6],
-            "filled_qty": row[7],
-            "avg_fill_price": row[8],
-            "shortfall_notional": row[9],
-            "execution_audit_json": row[10],
-            "raw_order": row[11],
-            "account_label": row[12],
-            "strategy_id": row[13],
-            "decision_hash": row[14],
-            "decision_json": row[15],
-            "alpaca_order_id": row[16],
-            "client_order_id": row[17],
-            "order_status": row[18],
-            "execution_tca_metric_id": str(row[19]) if row[19] is not None else None,
-        }
-    if len(row) >= 19:
-        return {
-            "execution_id": str(row[0]),
-            "trade_decision_id": str(row[1]),
-            "decision_created_at": row[2],
-            "computed_at": row[3],
-            "execution_event_at": row[3],
-            "execution_created_at": row[4],
-            "symbol": row[5],
-            "side": row[6],
-            "filled_qty": row[7],
-            "avg_fill_price": row[8],
-            "shortfall_notional": row[9],
-            "execution_audit_json": row[10],
-            "raw_order": row[11],
-            "account_label": row[12],
-            "strategy_id": row[13],
-            "decision_hash": row[14],
-            "decision_json": row[15],
-            "alpaca_order_id": row[16],
-            "client_order_id": row[17],
-            "order_status": row[18],
-        }
-    return {
-        "execution_id": None,
-        "trade_decision_id": str(row[12]),
-        "decision_created_at": row[0],
-        "computed_at": row[1],
-        "execution_event_at": row[1],
-        "execution_created_at": row[2],
-        "symbol": row[3],
-        "side": row[4],
-        "filled_qty": row[5],
-        "avg_fill_price": row[6],
-        "shortfall_notional": row[7],
-        "execution_audit_json": row[8],
-        "raw_order": row[9],
-        "account_label": row[10],
-        "strategy_id": row[11],
-        "decision_hash": row[12],
-        "decision_json": row[13],
-        "alpaca_order_id": row[14],
-        "client_order_id": row[15],
-        "order_status": row[16],
-    }
-
-
-def _execution_query_row_has_time(row: Sequence[object]) -> bool:
-    return (row[3] if len(row) >= 19 else row[1]) is not None
-
-
-def _execution_query_row_has_tca_ref(row: Mapping[str, object]) -> bool:
-    return (
-        _first_text(
-            row,
-            "execution_tca_metric_id",
-            "execution_tca_metric_ref",
-            "execution_tca_id",
-        )
-        is not None
-    )
-
-
-def _source_window_query_context(
-    row: Sequence[object],
-    *,
-    start_index: int,
-) -> dict[str, object]:
-    if len(row) < start_index + 15:
-        return {}
-    return {
-        "source_window_status": row[start_index],
-        "source_window_status_reason": row[start_index + 1],
-        "source_window_consumed_count": row[start_index + 2],
-        "source_window_inserted_count": row[start_index + 3],
-        "source_window_duplicate_count": row[start_index + 4],
-        "source_window_malformed_count": row[start_index + 5],
-        "source_window_missing_payload_count": row[start_index + 6],
-        "source_window_missing_identity_count": row[start_index + 7],
-        "source_window_out_of_scope_account_count": row[start_index + 8],
-        "source_window_unlinked_execution_count": row[start_index + 9],
-        "source_window_unlinked_decision_count": row[start_index + 10],
-        "source_window_failed_unhandled_count": row[start_index + 11],
-        "source_window_dropped_count": row[start_index + 12],
-        "source_window_gap_count": row[start_index + 13],
-        "source_window_gap_ranges": row[start_index + 14],
-    }
-
-
-def _order_lifecycle_query_row(row: Sequence[object]) -> dict[str, object]:
-    if len(row) >= 28:
-        return {
-            "execution_order_event_id": str(row[0]),
-            "trade_decision_id": str(row[1]),
-            "execution_id": str(row[2]) if row[2] is not None else None,
-            "event_ts": row[3],
-            "symbol": row[4],
-            "account_label": row[5],
-            "strategy_id": row[6],
-            "decision_hash": row[7],
-            "decision_json": row[8],
-            "alpaca_order_id": row[9],
-            "client_order_id": row[10],
-            "event_type": row[11],
-            "order_status": row[12],
-            "side": row[13],
-            "qty": row[14],
-            "filled_qty": row[15],
-            "filled_qty_delta": row[16],
-            "avg_fill_price": row[17],
-            "filled_notional_delta": row[18],
-            "fill_quantity_basis": row[19],
-            "event_fingerprint": row[20],
-            "source_topic": row[21],
-            "source_partition": row[22],
-            "source_offset": row[23],
-            "source_window_id": str(row[24]) if row[24] is not None else None,
-            "raw_event": row[25],
-            "execution_audit_json": row[26],
-            "raw_order": row[27],
-            **_source_window_query_context(row, start_index=28),
-        }
-    if len(row) >= 25:
-        return {
-            "execution_order_event_id": str(row[0]),
-            "trade_decision_id": str(row[1]),
-            "execution_id": str(row[2]) if row[2] is not None else None,
-            "event_ts": row[3],
-            "symbol": row[4],
-            "account_label": row[5],
-            "strategy_id": row[6],
-            "decision_hash": row[7],
-            "decision_json": row[8],
-            "alpaca_order_id": row[9],
-            "client_order_id": row[10],
-            "event_type": row[11],
-            "order_status": row[12],
-            "side": row[13],
-            "qty": row[14],
-            "filled_qty": row[15],
-            "avg_fill_price": row[16],
-            "event_fingerprint": row[17],
-            "source_topic": row[18],
-            "source_partition": row[19],
-            "source_offset": row[20],
-            "source_window_id": str(row[21]) if row[21] is not None else None,
-            "raw_event": row[22],
-            "execution_audit_json": row[23],
-            "raw_order": row[24],
-        }
-    if len(row) >= 21:
-        return {
-            "execution_order_event_id": str(row[0]),
-            "trade_decision_id": str(row[1]),
-            "execution_id": str(row[2]) if row[2] is not None else None,
-            "event_ts": row[3],
-            "symbol": row[4],
-            "account_label": row[5],
-            "strategy_id": row[6],
-            "decision_hash": row[7],
-            "decision_json": row[8],
-            "alpaca_order_id": row[9],
-            "client_order_id": row[10],
-            "event_type": row[11],
-            "order_status": row[12],
-            "event_fingerprint": row[13],
-            "source_topic": row[14],
-            "source_partition": row[15],
-            "source_offset": row[16],
-            "source_window_id": str(row[17]) if row[17] is not None else None,
-            "raw_event": row[18],
-            "execution_audit_json": row[19],
-            "raw_order": row[20],
-        }
-    if len(row) >= 20:
-        return {
-            "execution_order_event_id": str(row[0]),
-            "trade_decision_id": str(row[1]),
-            "execution_id": str(row[2]) if row[2] is not None else None,
-            "event_ts": row[3],
-            "symbol": row[4],
-            "account_label": row[5],
-            "strategy_id": row[6],
-            "decision_hash": row[7],
-            "decision_json": row[8],
-            "alpaca_order_id": row[9],
-            "client_order_id": row[10],
-            "event_type": row[11],
-            "order_status": row[12],
-            "event_fingerprint": row[13],
-            "source_topic": row[14],
-            "source_partition": row[15],
-            "source_offset": row[16],
-            "source_window_id": None,
-            "raw_event": row[17],
-            "execution_audit_json": row[18],
-            "raw_order": row[19],
-        }
-    return {
-        "execution_order_event_id": str(row[10]),
-        "trade_decision_id": str(row[4]),
-        "execution_id": None,
-        "event_ts": row[0],
-        "symbol": row[1],
-        "account_label": row[2],
-        "strategy_id": row[3],
-        "decision_hash": row[4],
-        "decision_json": row[5],
-        "alpaca_order_id": row[6],
-        "client_order_id": row[7],
-        "event_type": row[8],
-        "order_status": row[9],
-        "event_fingerprint": row[10],
-        "source_topic": row[11],
-        "source_partition": row[12],
-        "source_offset": row[13],
-        "source_window_id": None,
-        "raw_event": row[14],
-        "execution_audit_json": row[15],
-        "raw_order": row[16],
-    }
-
-
-def _order_lifecycle_query_row_has_time(row: Sequence[object]) -> bool:
-    return (row[3] if len(row) >= 20 else row[0]) is not None
-
-
-def _mark_runtime_ledger_tca_rows_as_exact_replay_artifacts(
-    tca_rows: list[dict[str, object]],
-) -> None:
-    for row in tca_rows:
-        row["authoritative"] = False
-        row["authority_reason"] = "exact_replay_artifact_not_runtime_proof"
-        row["promotion_blocker"] = "exact_replay_artifact_not_runtime_proof"
-        row["pnl_derivation"] = EXACT_REPLAY_ARTIFACT_AUTHORITY_CLASS
-        row["post_cost_promotion_eligible"] = False
-        blockers = _metadata_text_list(row.get("runtime_ledger_blockers"))
-        for blocker in EXACT_REPLAY_ARTIFACT_AUTHORITY_BLOCKERS:
-            if blocker not in blockers:
-                blockers.append(blocker)
-        row["runtime_ledger_blockers"] = blockers
-        bucket = row.get("runtime_ledger_bucket")
-        if not isinstance(bucket, Mapping):
-            continue
-        bucket_payload = {str(key): item for key, item in bucket.items()}
-        bucket_blockers = _metadata_text_list(bucket_payload.get("blockers"))
-        for blocker in EXACT_REPLAY_ARTIFACT_AUTHORITY_BLOCKERS:
-            if blocker not in bucket_blockers:
-                bucket_blockers.append(blocker)
-        bucket_payload["blockers"] = bucket_blockers
-        bucket_payload["authoritative"] = False
-        bucket_payload["authority_reason"] = "exact_replay_artifact_not_runtime_proof"
-        bucket_payload["pnl_derivation"] = EXACT_REPLAY_ARTIFACT_AUTHORITY_CLASS
-        bucket_payload["source_materialization"] = "exact_replay_artifact"
-        row["runtime_ledger_bucket"] = bucket_payload
-
-
-def _nonnegative_int(value: Any) -> int:
-    try:
-        return max(0, int(Decimal(str(value))))
-    except Exception:
-        return 0
-
-
-def _metadata_text_list(value: Any) -> list[str]:
-    if isinstance(value, str):
-        text = value.strip()
-        return [text] if text else []
-    if not isinstance(value, Sequence) or isinstance(value, (bytes, bytearray)):
-        return []
-    return [
-        text for item in cast(Sequence[Any], value) if (text := str(item or "").strip())
-    ]
-
-
-def _runtime_ledger_tca_ref_texts(value: Any) -> list[str]:
-    if isinstance(value, Mapping):
-        mapping = cast(Mapping[object, object], value)
-        for key in _EXECUTION_TCA_REF_KEYS + ("id", "ref"):
-            if (text := _text_or_none(mapping.get(key))) is not None:
-                return [text]
-        return []
-    if isinstance(value, Sequence) and not isinstance(value, (str, bytes, bytearray)):
-        refs: list[str] = []
-        for item in cast(Sequence[object], value):
-            if isinstance(item, Mapping):
-                refs.extend(_runtime_ledger_tca_ref_texts(item))
-                continue
-            if (text := _text_or_none(item)) is not None:
-                refs.append(text)
-        return refs
-    text = _text_or_none(value)
-    return [text] if text is not None else []
-
-
-def _runtime_ledger_execution_tca_metric_refs(
-    bucket: Mapping[str, object],
-) -> list[str]:
-    refs: list[str] = []
-    for key in _EXECUTION_TCA_REF_KEYS:
-        refs.extend(_runtime_ledger_tca_ref_texts(bucket.get(key)))
-    return list(dict.fromkeys(refs))
-
-
-def _metadata_symbol_list(value: Any) -> list[str]:
-    symbols: list[str] = []
-    for item in _metadata_text_list(value):
-        symbol = item.strip().upper()
-        if symbol:
-            symbols.append(symbol)
-    return list(dict.fromkeys(symbols))
-
-
-def _target_metadata_source_symbols(target_metadata: Mapping[str, Any]) -> list[str]:
-    return _metadata_symbol_list(target_metadata.get("paper_route_probe_symbols"))
-
-
-def _runtime_ledger_target_metadata_artifact_refs(
-    target_metadata: Mapping[str, Any],
-) -> list[str]:
-    refs: list[str] = []
-    for key in (
-        "runtime_ledger_artifact_refs",
-        "exact_replay_ledger_artifact_refs",
-    ):
-        refs.extend(_metadata_text_list(target_metadata.get(key)))
-    for key in (
-        "runtime_ledger_artifact_ref",
-        "exact_replay_ledger_artifact_ref",
-    ):
-        refs.extend(_metadata_text_list(target_metadata.get(key)))
-    return list(dict.fromkeys(refs))
-
-
-def _metadata_nonnegative_int_or_none(value: Any) -> int | None:
-    if value is None:
-        return None
-    text = str(value).strip()
-    if not text:
-        return None
-    return _nonnegative_int(value)
-
-
-def _runtime_ledger_target_metadata_blockers(
-    *,
-    target_metadata: Mapping[str, Any],
-    runtime_ledger_artifact_metadata: Mapping[str, Any],
-    window_start: datetime,
-    window_end: datetime,
-) -> list[str]:
-    metadata = _as_mapping(target_metadata)
-    if not metadata:
-        return []
-
-    blockers: list[str] = []
-    planned_refs = _runtime_ledger_target_metadata_artifact_refs(metadata)
-    loaded_refs = _metadata_text_list(
-        runtime_ledger_artifact_metadata.get("runtime_ledger_artifact_refs")
-    )
-    if loaded_refs and not _text_or_none(
-        runtime_ledger_artifact_metadata.get("runtime_ledger_artifact_authority_class")
-    ):
-        blockers.append("runtime_ledger_artifact_authority_class_missing")
-    if planned_refs:
-        if set(planned_refs) != set(loaded_refs):
-            blockers.append("runtime_ledger_artifact_refs_mismatch")
-
-    for keys, blocker in (
-        (
-            (
-                "runtime_ledger_artifact_row_count",
-                "exact_replay_ledger_artifact_row_count",
-            ),
-            "runtime_ledger_artifact_row_count_mismatch",
-        ),
-        (
-            (
-                "runtime_ledger_artifact_fill_count",
-                "exact_replay_ledger_artifact_fill_count",
-            ),
-            "runtime_ledger_artifact_fill_count_mismatch",
-        ),
-    ):
-        planned_count: int | None = None
-        for key in keys:
-            count = _metadata_nonnegative_int_or_none(metadata.get(key))
-            if count is not None:
-                planned_count = count
-                break
-        if planned_count is None:
-            continue
-        loaded_count = _nonnegative_int(runtime_ledger_artifact_metadata.get(keys[0]))
-        if planned_count != loaded_count:
-            blockers.append(blocker)
-
-    for key, actual_bound in (
-        ("window_start", window_start),
-        ("window_end", window_end),
-    ):
-        expected_bound = _parse_dt_or_none(metadata.get(key))
-        if metadata.get(key) is not None and expected_bound != actual_bound:
-            blockers.append("runtime_ledger_window_bounds_mismatch")
-            break
-
-    expected_candidate_id = _text_or_none(metadata.get("candidate_id"))
-    artifact_metadata_expected = bool(
-        planned_refs
-        or loaded_refs
-        or any(
-            metadata.get(key) is not None
-            for key in (
-                "runtime_ledger_artifact_row_count",
-                "exact_replay_ledger_artifact_row_count",
-                "runtime_ledger_artifact_fill_count",
-                "exact_replay_ledger_artifact_fill_count",
-                "replay_window_weekday_count",
-                "replay_min_window_weekday_count",
-            )
-        )
-    )
-    loaded_candidate_ids = _metadata_text_list(
-        runtime_ledger_artifact_metadata.get("runtime_ledger_artifact_candidate_ids")
-    )
-    loaded_candidate_id = _text_or_none(
-        runtime_ledger_artifact_metadata.get("runtime_ledger_artifact_candidate_id")
-    )
-    if expected_candidate_id is not None and artifact_metadata_expected:
-        if loaded_candidate_ids:
-            if expected_candidate_id not in loaded_candidate_ids:
-                blockers.append("runtime_ledger_artifact_candidate_id_mismatch")
-            elif len(loaded_candidate_ids) > 1:
-                blockers.append("runtime_ledger_artifact_candidate_id_ambiguous")
-        elif loaded_candidate_id is None:
-            blockers.append("runtime_ledger_artifact_candidate_id_missing")
-        elif expected_candidate_id != loaded_candidate_id:
-            blockers.append("runtime_ledger_artifact_candidate_id_mismatch")
-
-    planned_window_weekdays = _metadata_nonnegative_int_or_none(
-        metadata.get("replay_window_weekday_count")
-    )
-    loaded_window_weekdays = _metadata_nonnegative_int_or_none(
-        runtime_ledger_artifact_metadata.get(
-            "runtime_ledger_artifact_window_weekday_count"
-        )
-    )
-    if planned_window_weekdays is not None:
-        if loaded_window_weekdays is None:
-            blockers.append("runtime_ledger_artifact_window_weekday_count_missing")
-        elif planned_window_weekdays != loaded_window_weekdays:
-            blockers.append("runtime_ledger_artifact_window_weekday_count_mismatch")
-
-    min_window_weekdays = _metadata_nonnegative_int_or_none(
-        metadata.get("replay_min_window_weekday_count")
-    )
-    if min_window_weekdays is not None:
-        if loaded_window_weekdays is None:
-            blockers.append("runtime_ledger_artifact_window_weekday_count_missing")
-        elif loaded_window_weekdays < min_window_weekdays:
-            blockers.append("runtime_ledger_artifact_window_weekday_count_below_min")
-
-    return list(dict.fromkeys(blockers))
-
-
-def _runtime_window_source_kind_is_informational(
-    *,
-    source_kind: str,
-    target_metadata: Mapping[str, Any],
-) -> bool:
-    normalized = source_kind.strip().lower().replace("-", "_")
-    if normalized.startswith("simulation_"):
-        return True
-    if any(
-        marker in normalized
-        for marker in (
-            "informational",
-            "non_authoritative",
-            "offline_replay_triage",
-            "selection_only",
-        )
-    ):
-        return True
-    return False
-
-
-_AUTHORITATIVE_RUNTIME_LEDGER_MATERIALIZATION_SOURCE_KINDS = frozenset(
-    {
-        "paper_route_probe_runtime_observed",
-        "paper_runtime_observed",
-        "runtime_ledger_source_collection_candidate",
-        "live_runtime_observed",
-    }
-)
-
-
-def _source_collection_target_authorization_blockers(
-    target_metadata: Mapping[str, Any],
-) -> list[str]:
-    blockers: list[str] = []
-    if target_metadata.get("source_collection_authorized") is not True:
-        blockers.append("source_collection_authorization_missing")
-    if (
-        str(target_metadata.get("source_collection_authorization_scope") or "").strip()
-        != "source_window_evidence_collection_only"
-    ):
-        blockers.append("source_collection_authorization_scope_invalid")
-    return blockers
-
-
-def _source_kind_allows_runtime_ledger_materialization(
-    *,
-    source_kind: str,
-    target_metadata: Mapping[str, Any],
-) -> bool:
-    if _runtime_window_source_kind_is_informational(
-        source_kind=source_kind,
-        target_metadata=target_metadata,
-    ):
-        return False
-    normalized = source_kind.strip().lower().replace("-", "_")
-    if normalized == "runtime_ledger_source_collection_candidate":
-        return not _source_collection_target_authorization_blockers(target_metadata)
-    return normalized in _AUTHORITATIVE_RUNTIME_LEDGER_MATERIALIZATION_SOURCE_KINDS
-
-
-def _runtime_window_import_proof_hygiene_blockers(
-    *,
-    source_kind: str,
-    target_metadata: Mapping[str, Any],
-    dependency_quorum_decision: str,
-    continuity_ok: str,
-    drift_ok: str,
-) -> list[str]:
-    blockers: list[str] = []
-    normalized_source_kind = source_kind.strip().lower().replace("-", "_")
-    if not target_metadata and not _runtime_window_source_kind_is_informational(
-        source_kind=source_kind,
-        target_metadata=target_metadata,
-    ):
-        blockers.append("runtime_window_target_metadata_missing")
-    source_collection_candidate = (
-        normalized_source_kind == "runtime_ledger_source_collection_candidate"
-    )
-    if source_collection_candidate:
-        blockers.extend(
-            _source_collection_target_authorization_blockers(target_metadata)
-        )
-    if not source_collection_candidate and not dependency_quorum_decision.strip():
-        blockers.append("dependency_quorum_decision_missing")
-    if not source_collection_candidate and not continuity_ok.strip():
-        blockers.append("continuity_gate_missing")
-    if not source_collection_candidate and not drift_ok.strip():
-        blockers.append("drift_gate_missing")
-    for key in (
-        "runtime_ledger_target_metadata_blockers",
-        "runtime_window_import_health_gate_blockers",
-        "runtime_window_import_promotion_blockers",
-        "paper_route_session_import_blockers",
-        "candidate_blockers",
-        "runtime_window_import_audit_blockers",
-        "runtime_window_import_audit_target_blockers",
-    ):
-        blockers.extend(_metadata_text_list(target_metadata.get(key)))
-    return blockers
-
-
-def _runtime_ledger_profit_proof_present(
-    tca_rows: list[dict[str, object]],
-) -> bool:
-    for row in tca_rows:
-        if row.get("post_cost_expectancy_basis") != POST_COST_BASIS_RUNTIME_LEDGER:
-            continue
-        bucket = row.get("runtime_ledger_bucket")
-        if not isinstance(bucket, Mapping):
-            continue
-        if _runtime_ledger_bucket_profit_proof_present(bucket):
-            return True
-    return False
-
-
-def _runtime_observation_authority_payload(
-    *,
-    source_kind: str,
-    tca_rows: list[dict[str, object]],
-) -> dict[str, object]:
-    has_runtime_ledger_profit_proof = _runtime_ledger_profit_proof_present(tca_rows)
-    exact_replay_artifact_only = any(
-        "exact_replay_artifact_not_runtime_proof"
-        in _metadata_text_list(row.get("runtime_ledger_blockers"))
-        or row.get("authority_reason") == "exact_replay_artifact_not_runtime_proof"
-        for row in tca_rows
-    )
-    if has_runtime_ledger_profit_proof:
-        return {
-            "authoritative": True,
-            "authority_reason": "runtime_ledger_profit_proof",
-            "promotion_authority": "runtime_ledger",
-            "runtime_ledger_profit_proof_present": True,
-        }
-    if source_kind.startswith("simulation_"):
-        return {
-            "authoritative": False,
-            "authority_reason": "simulation_source_replay_only",
-            "promotion_authority": "blocked",
-            "runtime_ledger_profit_proof_present": has_runtime_ledger_profit_proof,
-        }
-    if exact_replay_artifact_only:
-        return {
-            "authoritative": False,
-            "authority_reason": "exact_replay_artifact_not_runtime_proof",
-            "promotion_authority": "blocked",
-            "runtime_ledger_profit_proof_present": False,
-        }
-    return {
-        "authoritative": False,
-        "authority_reason": "runtime_without_runtime_ledger_profit_proof",
-        "promotion_authority": "blocked",
-        "runtime_ledger_profit_proof_present": False,
-    }
-
-
-def _runtime_ledger_tca_materialization_metadata(
-    tca_rows: Sequence[Mapping[str, object]],
-) -> dict[str, object]:
-    bucket_row_count = 0
-    profit_proof_count = 0
-    authoritative_bucket_count = 0
-    source_execution_materialized_count = 0
-    execution_reconstruction_count = 0
-    target_notional_sizing_required_count = 0
-    target_notional_sizing_authoritative_count = 0
-    target_notional_sizing_missing_count = 0
-    target_notional_sizing_non_authoritative_count = 0
-    authority_reasons: list[str] = []
-    pnl_derivations: list[str] = []
-    source_materializations: list[str] = []
-    blockers: list[str] = []
-    profit_proof_blockers: list[str] = []
-    for row in tca_rows:
-        bucket = row.get("runtime_ledger_bucket")
-        if not isinstance(bucket, Mapping):
-            continue
-        bucket_payload = {str(key): value for key, value in bucket.items()}
-        bucket_row_count += 1
-        if _runtime_ledger_bucket_profit_proof_present(bucket_payload):
-            profit_proof_count += 1
-        else:
-            profit_proof_blockers.extend(
-                _runtime_ledger_bucket_profit_proof_blockers(bucket_payload)
-            )
-        if bool(row.get("authoritative")) or bool(bucket_payload.get("authoritative")):
-            authoritative_bucket_count += 1
-        authority_reason = _text_or_none(
-            row.get("authority_reason") or bucket_payload.get("authority_reason")
-        )
-        if authority_reason is not None:
-            authority_reasons.append(authority_reason)
-        pnl_derivation = _text_or_none(
-            row.get("pnl_derivation") or bucket_payload.get("pnl_derivation")
-        )
-        if pnl_derivation is not None:
-            pnl_derivations.append(pnl_derivation)
-        source_materialization = _text_or_none(
-            bucket_payload.get("source_materialization")
-        )
-        if source_materialization is not None:
-            source_materializations.append(source_materialization)
-        target_notional_sizing = _as_mapping(
-            row.get("paper_route_target_notional_sizing")
-            or bucket_payload.get("paper_route_target_notional_sizing")
-        )
-        if target_notional_sizing.get("requires_target_notional_sizing"):
-            target_notional_sizing_required_count += 1
-            target_notional_sizing_authoritative_count += _nonnegative_int(
-                target_notional_sizing.get("authoritative_target_notional_sizing_count")
-            )
-            target_notional_sizing_missing_count += _nonnegative_int(
-                target_notional_sizing.get("missing_target_notional_sizing_count")
-            )
-            target_notional_sizing_non_authoritative_count += sum(
-                _nonnegative_int(value)
-                for value in _as_mapping(
-                    target_notional_sizing.get("non_authoritative_sizing_source_counts")
-                ).values()
-            )
-            blockers.extend(_metadata_text_list(target_notional_sizing.get("blockers")))
-        if runtime_ledger_promotion_source_authority_present(bucket_payload):
-            source_execution_materialized_count += 1
-        if authority_reason == "execution_reconstruction_not_runtime_ledger_proof":
-            execution_reconstruction_count += 1
-        for blocker_source in (
-            row.get("runtime_ledger_blockers"),
-            row.get("promotion_blocker"),
-            bucket_payload.get("blockers"),
-        ):
-            blockers.extend(_metadata_text_list(blocker_source))
-    blockers.extend(profit_proof_blockers)
-    return {
-        "runtime_ledger_tca_row_count": len(tca_rows),
-        "runtime_ledger_tca_runtime_bucket_row_count": bucket_row_count,
-        "runtime_ledger_tca_profit_proof_count": profit_proof_count,
-        "runtime_ledger_tca_authoritative_bucket_count": authoritative_bucket_count,
-        "runtime_ledger_source_execution_materialized_bucket_count": (
-            source_execution_materialized_count
-        ),
-        "runtime_ledger_execution_reconstruction_bucket_count": (
-            execution_reconstruction_count
-        ),
-        "paper_route_target_notional_sizing_required_count": (
-            target_notional_sizing_required_count
-        ),
-        "paper_route_target_notional_sizing_authoritative_count": (
-            target_notional_sizing_authoritative_count
-        ),
-        "paper_route_target_notional_sizing_missing_count": (
-            target_notional_sizing_missing_count
-        ),
-        "paper_route_target_notional_sizing_non_authoritative_count": (
-            target_notional_sizing_non_authoritative_count
-        ),
-        "runtime_ledger_materialization_authority_reasons": sorted(
-            dict.fromkeys(authority_reasons)
-        ),
-        "runtime_ledger_materialization_pnl_derivations": sorted(
-            dict.fromkeys(pnl_derivations)
-        ),
-        "runtime_ledger_source_materializations": sorted(
-            dict.fromkeys(source_materializations)
-        ),
-        "runtime_ledger_materialization_blockers": sorted(dict.fromkeys(blockers)),
-        "runtime_ledger_profit_proof_blockers": sorted(
-            dict.fromkeys(profit_proof_blockers)
-        ),
-    }
-
-
-def _runtime_ledger_tca_rows_from_observed_buckets(
-    buckets: Sequence[object],
-) -> list[dict[str, object]]:
-    rows: list[dict[str, object]] = []
-    for bucket in buckets:
-        payload_json = _as_mapping(getattr(bucket, "payload_json", {}))
-        bucket_payloads: list[Any] = []
-        raw_payloads = _as_sequence(payload_json.get("runtime_ledger_buckets")) or []
-        bucket_payloads.extend(raw_payloads)
-        single_payload = _as_mapping(payload_json.get("runtime_ledger_bucket"))
-        if single_payload:
-            bucket_payloads.append(single_payload)
-        for raw_payload in bucket_payloads:
-            ledger_payload = _as_mapping(raw_payload)
-            if not ledger_payload:
-                continue
-            rows.append(
-                {
-                    "runtime_ledger_bucket": ledger_payload,
-                    "authoritative": ledger_payload.get("authoritative"),
-                    "authority_reason": ledger_payload.get("authority_reason"),
-                    "pnl_derivation": ledger_payload.get("pnl_derivation"),
-                    "runtime_ledger_blockers": ledger_payload.get("blockers"),
-                    "post_cost_expectancy_basis": ledger_payload.get("pnl_basis"),
-                    "paper_route_target_notional_sizing": ledger_payload.get(
-                        "paper_route_target_notional_sizing"
-                    ),
-                }
-            )
-    return rows
-
-
-def _runtime_observation_payload_with_observed_runtime_ledger_materialization(
-    *,
-    source_kind: str,
-    runtime_observation_payload: Mapping[str, Any],
-    buckets: Sequence[object],
-) -> dict[str, Any]:
-    observed_rows = _runtime_ledger_tca_rows_from_observed_buckets(buckets)
-    if not observed_rows:
-        return dict(runtime_observation_payload)
-
-    observed_materialization = _runtime_ledger_tca_materialization_metadata(
-        observed_rows
-    )
-    observed_authority = _runtime_observation_authority_payload(
-        source_kind=source_kind,
-        tca_rows=observed_rows,
-    )
-    observed_materialization_blockers = sorted(
-        dict.fromkeys(
-            [
-                *_metadata_text_list(
-                    observed_materialization.get(
-                        "runtime_ledger_materialization_blockers"
-                    )
-                ),
-                *_metadata_text_list(
-                    observed_materialization.get("runtime_ledger_profit_proof_blockers")
-                ),
-            ]
-        )
-    )
-    source_diagnostics = {
-        **_as_mapping(runtime_observation_payload.get("source_activity_diagnostics")),
-        "observed_runtime_ledger_materialization_authority": True,
-        "observed_runtime_ledger_tca_row_count": observed_materialization[
-            "runtime_ledger_tca_row_count"
-        ],
-        "observed_runtime_ledger_runtime_bucket_row_count": (
-            observed_materialization["runtime_ledger_tca_runtime_bucket_row_count"]
-        ),
-        "observed_runtime_ledger_profit_proof_count": observed_materialization[
-            "runtime_ledger_tca_profit_proof_count"
-        ],
-        "observed_runtime_ledger_materialization_blockers": (
-            observed_materialization["runtime_ledger_materialization_blockers"]
-        ),
-        "observed_runtime_ledger_profit_proof_blockers": (
-            observed_materialization["runtime_ledger_profit_proof_blockers"]
-        ),
-    }
-    return {
-        **dict(runtime_observation_payload),
-        **observed_authority,
-        **observed_materialization,
-        "runtime_ledger_materialization_authority_source": ("observed_runtime_buckets"),
-        "source_activity_diagnostics": source_diagnostics,
-        "source_activity_diagnostic_blockers": observed_materialization_blockers,
-        "source_activity_diagnostic_blockers_from_source_query": (
-            _metadata_text_list(
-                runtime_observation_payload.get("source_activity_diagnostic_blockers")
-            )
-        ),
-    }
-
-
-def _runtime_window_import_audit_summary(
-    *,
-    run_id: str,
-    candidate_id: str | None,
-    hypothesis_id: str,
-    observed_stage: str,
-    strategy_name: str,
-    strategy_names: Sequence[str],
-    account_label: str,
-    source_account_label: str,
-    window_start: datetime,
-    window_end: datetime,
-    source_kind: str,
-    source_manifest_ref: str | None,
-    dataset_snapshot_ref: str | None,
-    decisions: Sequence[datetime],
-    executions: Sequence[datetime],
-    tca_rows: Sequence[Mapping[str, object]],
-    buckets: Sequence[object],
-    runtime_observation_payload: Mapping[str, Any],
-    runtime_ledger_target_metadata_blockers: Sequence[str],
-) -> dict[str, Any]:
-    ledger_payloads = [
-        payload
-        for row in tca_rows
-        if (payload := _as_mapping(row.get("runtime_ledger_bucket")))
-    ]
-    ledger_profit_proof_blockers = sorted(
-        {
-            blocker
-            for payload in ledger_payloads
-            for blocker in _runtime_ledger_bucket_profit_proof_blockers(payload)
-        }
-    )
-    observed_bucket_payloads = [
-        _as_mapping(getattr(bucket, "payload_json", {})) for bucket in buckets
-    ]
-    observed_bucket_blockers = sorted(
-        {
-            blocker
-            for payload in observed_bucket_payloads
-            for blocker in _metadata_text_list(
-                payload.get("runtime_ledger_profit_proof_blockers")
-                or payload.get("promotion_blocking_reasons")
-                or payload.get("blockers")
-            )
-        }
-    )
-    source_activity_diagnostics = _as_mapping(
-        runtime_observation_payload.get("source_activity_diagnostics")
-    )
-    return {
-        "schema_version": "torghut.runtime-window-import-source-audit.v1",
-        "audit_only": True,
-        "would_persist": False,
-        "verdict": (
-            "profit_proof_present"
-            if _runtime_ledger_profit_proof_present(tca_rows)
-            else "blocked"
-        ),
-        "run_id": run_id,
-        "candidate_id": candidate_id,
-        "hypothesis_id": hypothesis_id,
-        "observed_stage": observed_stage,
-        "strategy_name": strategy_name,
-        "strategy_name_candidates": list(strategy_names),
-        "account_label": account_label,
-        "source_account_label": source_account_label,
-        "window_start": window_start.isoformat(),
-        "window_end": window_end.isoformat(),
-        "source_kind": source_kind,
-        "source_manifest_ref": source_manifest_ref,
-        "dataset_snapshot_ref": dataset_snapshot_ref,
-        "decision_count": len(decisions),
-        "execution_count": len(executions),
-        "tca_row_count": len(tca_rows),
-        "runtime_ledger_bucket_row_count": len(ledger_payloads),
-        "observed_runtime_bucket_count": len(buckets),
-        "observed_runtime_bucket_blockers": observed_bucket_blockers,
-        "runtime_ledger_profit_proof_present": _runtime_ledger_profit_proof_present(
-            tca_rows
-        ),
-        "runtime_ledger_profit_proof_blockers": ledger_profit_proof_blockers,
-        "runtime_ledger_target_metadata_blockers": list(
-            dict.fromkeys(runtime_ledger_target_metadata_blockers)
-        ),
-        "source_activity_diagnostics": source_activity_diagnostics,
-        "source_activity_diagnostic_blockers": _metadata_text_list(
-            runtime_observation_payload.get("source_activity_diagnostic_blockers")
-        ),
-        "promotion_authority": runtime_observation_payload.get("promotion_authority"),
-        "authority_reason": runtime_observation_payload.get("authority_reason"),
-        "runtime_observation": {
-            "runtime_ledger_profit_proof_present": runtime_observation_payload.get(
-                "runtime_ledger_profit_proof_present"
-            ),
-            "evidence_provenance": runtime_observation_payload.get(
-                "evidence_provenance"
-            ),
-            "source_kind": runtime_observation_payload.get("source_kind"),
-            "source_activity_symbol_filter": runtime_observation_payload.get(
-                "source_activity_symbol_filter"
-            ),
-        },
-        "runtime_ledger_materialization": _runtime_ledger_tca_materialization_metadata(
-            tca_rows
-        ),
-    }
-
-
 def _strategy_name_candidates(*values: str | None) -> list[str]:
     candidates: list[str] = []
     for value in values:
@@ -3213,18 +348,6 @@ def _strategy_name_candidates(*values: str | None) -> list[str]:
             if normalized and normalized not in candidates:
                 candidates.append(normalized)
     return candidates
-
-
-def _execution_signed_qty(*, side: Any, qty: Any) -> Decimal:
-    normalized_side = str(side or "").strip().lower().replace("-", "_")
-    quantity = _decimal_or_none(qty)
-    if quantity is None or quantity <= 0:
-        return Decimal("0")
-    if normalized_side in {"buy", "buy_to_cover", "cover"}:
-        return quantity
-    if normalized_side in {"sell", "sell_short", "short"}:
-        return -quantity
-    return Decimal("0")
 
 
 def _runtime_execution_cost_amount(
@@ -3599,31 +722,6 @@ def _with_linked_execution_id(
     return payload
 
 
-def _execution_row_has_fill(row: Mapping[str, object]) -> bool:
-    filled_qty = _first_positive_decimal(
-        row,
-        "filled_qty",
-        "filled_quantity",
-        "qty",
-        "quantity",
-    )
-    avg_fill_price = _first_positive_decimal(
-        row,
-        "avg_fill_price",
-        "filled_avg_price",
-        "filled_average_price",
-        "average_fill_price",
-        "fill_price",
-        "filled_price",
-    )
-    return (
-        filled_qty is not None
-        and filled_qty > 0
-        and avg_fill_price is not None
-        and avg_fill_price > 0
-    )
-
-
 def _runtime_lifecycle_identifier_values(row: Mapping[str, object]) -> set[str]:
     values = _text_values(row, *_RUNTIME_LIFECYCLE_IDENTIFIER_KEYS)
     return {value for value in values if value}
@@ -3721,17 +819,6 @@ def _order_feed_fill_lifecycle_blockers(
     return blockers
 
 
-def _source_authority_order_event_row(row: Mapping[str, object]) -> bool:
-    return any(
-        row.get(key) is not None
-        for key in (
-            "execution_order_event_id",
-            "source_window_id",
-            "source_offset",
-        )
-    )
-
-
 def _fill_quantity_basis(row: Mapping[str, object]) -> str | None:
     raw = _first_text(row, "fill_quantity_basis", "quantity_basis")
     if raw is None:
@@ -3811,11 +898,6 @@ def _order_feed_fill_delta_blockers(row: Mapping[str, object]) -> list[str]:
     ):
         return [ORDER_FEED_FILL_DELTA_MISSING_BLOCKER]
     return []
-
-
-def _runtime_source_row_symbol(row: Mapping[str, object]) -> str | None:
-    symbol = _first_text(row, "symbol", "ticker")
-    return symbol.upper() if symbol is not None else None
 
 
 def _runtime_source_row_in_bucket(
@@ -5692,47 +2774,6 @@ def _runtime_ledger_artifact_rows(payload: Mapping[str, Any]) -> list[dict[str, 
     return []
 
 
-def _runtime_ledger_event_type(row: Mapping[str, Any]) -> str:
-    raw = _text_or_none(
-        row.get("ledger_event_type")
-        or row.get("runtime_ledger_event_type")
-        or row.get("lifecycle_event")
-        or row.get("event_type")
-        or row.get("order_event_type")
-        or row.get("order_status")
-        or row.get("status")
-    )
-    if raw is None:
-        if any(row.get(key) is not None for key in ("filled_qty", "qty", "quantity")):
-            return "fill"
-        if any(row.get(key) is not None for key in ("order_id", "alpaca_order_id")):
-            return "order_submitted"
-        if any(
-            row.get(key) is not None
-            for key in ("decision_id", "trade_decision_id", "decision_hash")
-        ):
-            return "decision"
-        return "diagnostic"
-    normalized = raw.lower().replace("-", "_").replace(" ", "_")
-    return {
-        "trade_decision": "decision",
-        "signal_decision": "decision",
-        "new_order": "order_submitted",
-        "submitted": "order_submitted",
-        "accepted": "order_submitted",
-        "new": "order_submitted",
-        "filled": "fill",
-        "partially_filled": "partial_fill",
-    }.get(normalized, normalized)
-
-
-def _runtime_ledger_row_time(row: Mapping[str, Any]) -> datetime | None:
-    for key in ("executed_at", "filled_at", "event_ts", "created_at", "computed_at"):
-        if (parsed := _parse_dt_or_none(row.get(key))) is not None:
-            return parsed
-    return None
-
-
 def _runtime_ledger_activity_times(
     rows: list[dict[str, Any]],
 ) -> tuple[list[datetime], list[datetime]]:
@@ -5748,30 +2789,6 @@ def _runtime_ledger_activity_times(
         elif event_type in _RUNTIME_LEDGER_FILL_EVENTS:
             execution_times.append(event_time)
     return decision_times, execution_times
-
-
-def _parse_artifact_window_datetime(
-    value: Any,
-    *,
-    date_end: bool = False,
-) -> datetime | None:
-    if isinstance(value, datetime):
-        return (
-            value.astimezone(timezone.utc)
-            if value.tzinfo
-            else value.replace(tzinfo=timezone.utc)
-        )
-    text = str(value or "").strip()
-    if not text:
-        return None
-    try:
-        if "T" not in text and len(text) == 10:
-            parsed_date = date.fromisoformat(text)
-            parsed = datetime.combine(parsed_date, time.min, tzinfo=timezone.utc)
-            return parsed + timedelta(days=1) if date_end else parsed
-        return _parse_dt(text)
-    except Exception:
-        return None
 
 
 def _runtime_ledger_artifact_window_metadata(
@@ -5814,43 +2831,6 @@ def _runtime_ledger_artifact_window_metadata(
             end=end,
         ),
     }
-
-
-def _window_weekday_count(*, start: datetime, end: datetime) -> int:
-    count = 0
-    current = datetime.combine(start.date(), time.min, tzinfo=timezone.utc)
-    while current < end:
-        next_day = current + timedelta(days=1)
-        if current.weekday() < 5 and next_day > start:
-            count += 1
-        current = next_day
-    return count
-
-
-def _runtime_ledger_artifact_candidate_ids(
-    *,
-    payload: Mapping[str, Any],
-    rows: Sequence[Mapping[str, Any]],
-) -> list[str]:
-    candidates: list[str] = []
-    for key in (
-        "candidate_id",
-        "candidateId",
-        "strategy_candidate_id",
-        "strategyCandidateId",
-    ):
-        if candidate_id := _text_or_none(payload.get(key)):
-            candidates.append(candidate_id)
-    for row in rows:
-        for key in (
-            "candidate_id",
-            "candidateId",
-            "strategy_candidate_id",
-            "strategyCandidateId",
-        ):
-            if candidate_id := _text_or_none(row.get(key)):
-                candidates.append(candidate_id)
-    return sorted(dict.fromkeys(candidates))
 
 
 def _runtime_ledger_tca_rows_from_artifacts(
